+++ .husky/pre-commit
@@ -3,3 +3,6 @@
 
 cd frontend
 npx lint-staged
+cd ..
+cd backend
+npx lint-staged
