+++ backend/.lintstagedrc.json
@@ -1,0 +1,7 @@
+{
+  "app/**/*.{js,ts,jsx,tsx}": [
+    "eslint --cache --fix",
+    "./node_modules/.bin/eslint --max-warnings=0",
+    "./node_modules/.bin/prettier --write"
+  ]
+}
