+++ backend/.prettierrc
@@ -1,20 +1,5 @@
 {
-	"arrowParens": "always",
-	"bracketSpacing": false,
-	"endOfLine": "lf",
-	"htmlWhitespaceSensitivity": "css",
-	"insertPragma": false,
-	"jsxBracketSameLine": false,
-	"jsxSingleQuote": false,
 	"printWidth": 100,
-	"proseWrap": "preserve",
-	"quoteProps": "as-needed",
-	"requirePragma": false,
-	"semi": true,
-	"singleQuote": false,
-	"tabWidth": 4,
-	"trailingComma": "es5",
-	"useTabs": true,
-	"vueIndentScriptAndStyle": false,
-	"parser": "babel"
+	"semi": false,
+	"singleQuote": true
 }
