+++ backend/package.json
@@ -39,10 +39,13 @@
   "devDependencies": {
     "cross-env": "7.0.3",
     "ejs": "3.1.6",
-    "eslint": "7.28.0",
+    "eslint": "8.14.0",
+    "eslint-config-airbnb-base": "15.0.0",
     "eslint-config-prettier": "8.3.0",
+    "eslint-plugin-prettier": "4.0.0",
+    "lint-staged": "12.4.0",
     "nodemon": "2.0.7",
-    "pino-pretty": "^7.6.1",
+    "pino-pretty": "7.6.1",
     "prettier": "2.3.1",
     "prisma": "3.8.1"
   }
