+++ frontend/src/components/organisms/InteractiveMap/index.js
@@ -1,0 +1,4 @@
+import InteractiveMap from './InteractiveMap'
+import MapOfDistricts from './MapOfDistricts'
+
+export { MapOfDistricts, InteractiveMap }
