+++ frontend/src/components/organisms/InteractiveMap/store/context.js
@@ -1,0 +1,4 @@
+import { createContext } from 'react'
+
+const MapContext = createContext(null)
+export default MapContext
