+++ frontend/src/components/organisms/index.js
@@ -1,0 +1,4 @@
+import { MapOfDistricts } from './InteractiveMap'
+import Modal from './Modal/Modal'
+
+export { Modal, MapOfDistricts }
