--- frontend/src/components/organisms/interactive-map/InteractiveMap.jsx
@@ -1,2272 +1,0 @@
-import {useContext, useEffect, useState} from "react";
-import {MapContext} from "./store/context";
-import {MAP_ACTIONS} from "./store/reducer";
-import styled from "styled-components";
-
-const SvgStyled = styled.svg`
-	width: 100%;
-  
-	path {
-		cursor: pointer;
-		fill: transparent;
-		pointer-events: auto;
-	}
-
-	#Ciutat-Vella,
-	#Les-Corts,
-	#L-Eixample,
-	#Horta-Guinardo,
-	#Nou-Barris,
-	#Sants-Montjuic,
-	#Sant-Andreu,
-	#Sant-Marti,
-	#Gracia,
-	#Sarria-Sant-Gervasi,
-	#path25 {
-		fill: #fcf5e3;
-		cursor: pointer;
-	}
-
-	.lit-neighborhood {
-		color: #fff;
-		fill: #db2c7f !important;
-	}
-
-	.lit-district {
-		fill: #db2c7f !important;
-	}
-`;
-
-const InteractiveMap = () => {
-	const {state, dispatch} = useContext(MapContext);
-	const [screenWidth, setScreenWidth] = useState(window.innerWidth);
-
-	const handleWindowSizeChange = () => {
-		setScreenWidth(window.innerWidth);
-	};
-
-	useEffect(() => {
-		window.addEventListener("resize", handleWindowSizeChange);
-		return () => {
-			window.removeEventListener("resize", handleWindowSizeChange);
-		};
-	}, []);
-
-	return (
-		<div>
-			<SvgStyled
-				xmlns="http://www.w3.org/2000/svg"
-				xmlnsXlink="http://www.w3.org/1999/xlink"
-				width="427.543"
-				height="405.636"
-				viewBox="0 0 427.543 405.636"
-				onMouseOver={(e) => {
-					if (screenWidth <= 768) {
-						dispatch({type: MAP_ACTIONS.LIT_DISTRICT, payload: e.target.id});
-					} else dispatch({type: MAP_ACTIONS.LIT_NEIGHBORHOOD, payload: e.target.id});
-				}}
-			>
-				<defs>
-					<clipPath id="clip-path">
-						<path id="path693" d="M0,154.905H80.429V0H0Z" />
-					</clipPath>
-				</defs>
-				<g id="Group_309" data-name="Group 309" transform="translate(-20014.055 -3139.501)">
-					<g id="layer1" transform="translate(20014.555 3140.001)">
-						<path
-							id="Sarria-Sant-Gervasi"
-							d="M275.091,155.246c.084-.667.861-.884,1.281-1.332.881-.857,1.308.114,1.913.8a1.973,1.973,0,0,0,3.377-.725,18.5,18.5,0,0,1,.717-3.725,5.3,5.3,0,0,0,.91-1.372c.324-.428.162-.963.109-1.446.489-.458,1.374-.517,1.46-1.306-.2-.417-.781-.736-1.129-1.114-1.188-.672.781-1.782.347-2.859-.155-1.514,1.639-2.211,2.554-3.127.53-.237,1.054-.619.806-1.279a15.91,15.91,0,0,1-.544-3.877,6.909,6.909,0,0,1,.519-2.218,1.668,1.668,0,0,1,1.211.554,7.826,7.826,0,0,0,2.787,3.518c.542.166.688.944,1.305.881a12.415,12.415,0,0,0,3.072,1.553c1.162,1.024,1.887-.386,2.8-.9a4.021,4.021,0,0,0,1.073-.934c.352.117.9.777,1.177.311-.556-.9-1.531-1.572-2.029-2.552-.885-1.546-1.27-3.453-2.562-4.729a8.8,8.8,0,0,0-2.192-1.752c-.419-.516-.687-1.215-1.425-1.334-1.353-.342-1.856-1.842-2.549-2.9-.811-.819-.7-1.431.248-1.919.638-.432,1.088-1.248,1.926-1.257a10.216,10.216,0,0,1,1.2-.933,19.653,19.653,0,0,0,1.812-1.178c1.6.446,3.253.7,4.806,1.279,1.163.5,2.309,1.042,3.465,1.559a17.938,17.938,0,0,1,.112-3.18c.039-1.249-.785-2.371-.724-3.633a5.174,5.174,0,0,0-.572-2.366c-.437-1.968,1.222-3.547,1.329-5.454.4-1.453.559-3.016,1.463-4.272a26.97,26.97,0,0,1,1.735-2.757,5.413,5.413,0,0,1,2.139-1.466c1.076-.408,2.135-.92,3.221-1.265a16.507,16.507,0,0,1,3.815.5c.852.06-.4.464-.647.7l-3.947,2.357c-.161.549-.214,1.052.429,1.258a16.277,16.277,0,0,0,3.2,1.58c.256.755-.388,1.341-.652,2-.33.492-.469,1.17-.918,1.546-.849.553-1.932.838-2.548,1.682-.232.746-.449,1.716.294,2.252.592.576,1.334,1.045,1.538,1.9.22.76-.015,1.585.137,2.37a3.794,3.794,0,0,0,1.186,2.458,23.015,23.015,0,0,1,1.841,2.46c.2.993-.8,1.689-1.57,2.095-.555.587.005,1.446.347,1.988.377.443.965.51,1.137-.144.292-.517.8-1.106,1.458-.84a1.883,1.883,0,0,1,2.688.533c.24.584-.019,1.432-.732,1.475-.747.108-1.106.879-1.372,1.493.343,1.107,2.876,1.027,2.22,2.607-.559.3-1.485-.058-1.833.626-.192.9.292,1.782.412,2.67a14.187,14.187,0,0,1,.487,2.57.7.7,0,0,0,.219,1.169,4.684,4.684,0,0,1,1.593,2.018c.282.38.607.821.321,1.294-.092.651-.318,1.392.151,1.957a6.348,6.348,0,0,0,2.127-.125,16.708,16.708,0,0,0,1.611-.463c.757.894.056,3.056,1.227,3.277.821.225,1.341-.5,2.026-.128.237.461.872-.134,1.311.223.451.121.911.686,1.364.519.618-.289,1.186-.984,1.929-.742a10.763,10.763,0,0,0,1.435-.439,14.411,14.411,0,0,0,2.175-.111,32.841,32.841,0,0,0,3.557-1.215c.777-.08,1.511-.361,2.163.19.444.368.945.182,1.334.612a6.255,6.255,0,0,1,1.275,1.412,3.242,3.242,0,0,0,.035,2.2c.258.578.077,1.255.484,1.776a13.667,13.667,0,0,1,.706,2.437c.064.565-.491.736-.887.991-.5.35-1.184.532-1.313,1.206-.291.543-.458,1.175-.774,1.683.926.43,1.327,1.729,1.744,2.716.21.549.724,1.1.624,1.7-.381.915-.209,1.934-.532,2.872-.164.482-.4.9-.114,1.394-.071.5.426.724.511,1.195.114.462-.248,1.187.22,1.452a6.244,6.244,0,0,0,1.9.487c1.007.5,2.092.9,3.052,1.453.4.361.456.775.793,1.189a12.262,12.262,0,0,0,3.421,3.474,17.913,17.913,0,0,1,2,2.43c.664,1.085,1.268,2.221,1.969,3.274.519.354.913.951,1.526,1.121.507-.243,1.276-.049,1.288.6.123.457-.013,1.159.625,1.22,1.26.516.867,2.432,1.619,3.439.878.789.456,1.89-.031,2.755a3.105,3.105,0,0,0-.322,2.8c.119.532.346,1.016.419,1.544a2.686,2.686,0,0,0,1.107,1.548,12.338,12.338,0,0,1,1.371,3.14c.283.813.455,1.755,1.272,2.214a7.338,7.338,0,0,0,1.682,1.2c.567.033.341.8-.027,1.03-1.3.656-2.745,1.037-4.1,1.581-1.32.413-1.816,1.848-2.8,2.666-.891.261-.872,1.358-1.329,2.021-.639.951-.276,2.15-.744,3.162-.16.52-.66,1.1-1.245.765-.748-.428-.229.475-.2.87.258.27.307,1.01.485,1.462.66,2.274,1.217,4.587,1.942,6.837.755,2.04,1.137,4.2,1.719,6.294l.976,3.853c.366.538.376-.439.417-.828.241-.841-.332-1.717,0-2.529.352-.38,1.055.035,1.363-.487a2.284,2.284,0,0,0-.12-1.965c-.007-.538.8-.685,1.2-.409.428.187,1.081-.012,1.2.609.33.425,1.2-.134,1.334.5-.149.488-.447,1.02-.106,1.517-.066,1.075,1.123.523,1.553.971-.2.643-1.119.789-1.375,1.429a4.913,4.913,0,0,0-.8,2.158,12.3,12.3,0,0,0,1.086,1.478,18.305,18.305,0,0,1,2.146.7,25.274,25.274,0,0,1,1.025,3.042,6.275,6.275,0,0,1,.544,2.766c-.055,2.157-.322,4.293-.509,6.441a11.565,11.565,0,0,0-1.7,1.089c-.1,1.288-.308,2.591-.341,3.871a3.248,3.248,0,0,1,.417,1.892c-.025,2.723-.05,5.447-.125,8.166.053.8-.137,1.718.552,2.309,1.576,2.28,3.448,4.34,5.08,6.573.693.874.152,2.018-.073,2.965a9.719,9.719,0,0,0-1.093,2.753c-.848,3.1-1.824,6.266-1.6,9.524-.7-.192-.862.205-1.193.415-.04.083-.027.883.221,1,.048.884.135.172.417-.009.732-.146.417.653.434,1.084-.12.977.181,2.2-.716,2.882-1.161.444-1.45,1.778-2.069,2.729a1.987,1.987,0,0,1-1.554,1.4,5.715,5.715,0,0,0-2.115,1.092c-.6,1.448-.881,3.006-1.35,4.5a3.393,3.393,0,0,1-1.038,2.15c-.5.623-1.525.955-1.426,1.92-.826,4.636-1.17,9.343-2.029,13.97a13.99,13.99,0,0,1-1.252,3.348c-7.77-3.91-15.574-7.752-23.36-11.628a1.014,1.014,0,0,1-1.211.727c-.473-.3-.9-.622-.793-1.241.086-.623-.736-.626-1.115-.932q-5.573-2.886-11.139-5.787c-1.356-2.519-3.057-4.829-4.587-7.243-1.959-2.884-3.818-5.839-5.843-8.676a3.349,3.349,0,0,0-2-1.5c-4.366-2.069-8.661-4.285-13.009-6.391-1.575-1.159-3.635-1.281-5.4-2.042.773-1.535,1.53-3.256.915-4.991-.384-1.77-1.954-3.163-1.849-5.045,3.3.081,6.827-2.5,6.761-6-.1-2.036-1.88-3.536-1.981-5.577-.38-2.166-.69-4.343-.974-6.523a5.43,5.43,0,0,0-2.394-.073,9.123,9.123,0,0,1-3.806-.351c-.64-.4-1.691-.783-1.518-1.719a2.233,2.233,0,0,1,.353-1.593,4.539,4.539,0,0,0,.685-1.939c.494-.3.421-.862.381-1.35.137-.676.712-.858.847-1.657.108-.61.687-.7.745-1.25-.069-.533-.331-1.091.069-1.557.012-.722.377-1.5-.069-2.146a1.3,1.3,0,0,1,.242-1.211,3.592,3.592,0,0,0,.035-1.384,2.627,2.627,0,0,0-.554-1.661,1.667,1.667,0,0,0-.346-1.28c.009-.426.168-.9-.3-1.241-.992-1.184-1.049-2.922-2.111-4.048-1.128-.573.054-.719.566-1.091.92-.55,1.628-1.493,2.745-1.679a5.54,5.54,0,0,0,3.747-2.807,3.922,3.922,0,0,0,.61-1.671,14.266,14.266,0,0,0-.985-2.528,1.09,1.09,0,0,1-.183-1.38c.5-1.008,1.524-1.62,2.124-2.57.465-.523.191-1.367.809-1.8a2.614,2.614,0,0,0,.958-1.616,11.62,11.62,0,0,1,.946-2.862c-1.583-1.966-1.307-5.032-3.438-6.6-1.025-.516-1.956.493-2.988.485a5.145,5.145,0,0,1-1.719-.193c-.693-.421-1.527-.681-1.9-1.471-1.63-2.262-2.653-5.153-5.215-6.578-1.665-2.8-4.075-5.319-4.71-8.613.059-1.345.142-2.91-1.093-3.806a2.562,2.562,0,0,0-1.6-.846,27.2,27.2,0,0,0-3.69,0,21.515,21.515,0,0,0-5.28.5c-.059.7-.8.274-1.257.282-1.119-.189-2.211-.524-3.311-.767-.608.432-1.329-.153-1.938.1a8.774,8.774,0,0,1-2.15,2.139c-1.216.779-2.728.47-4.094.673a4.423,4.423,0,0,1-2.824-.216,1.4,1.4,0,0,1-1.7-.554c-.959-1.886-2.05-4.238-4.387-4.614-.916.107-.165-.97-.355-1.51a11.274,11.274,0,0,0-.173-2.216,10.258,10.258,0,0,1-1.579-.914l-.5-1.924a29.557,29.557,0,0,0,1.107-9.379l.934-1.523a1.566,1.566,0,0,1,1.419.524c.941.521,1.691,1.34,2.576,1.916a2.583,2.583,0,0,0,1.923-.294c.76-1.689,1.41-3.675,3.289-4.428,1.148-.633,2.325-1.06,2.936.4.951,1.576,2.824.93,4.31,1.194,1.567.256,3.521.815,3.842,2.627.127.71-.248,1.612.652.866.689-1.035,3.01.691,2.826-.622l-.308-2.187Zm-1.246-9.258a29.27,29.27,0,0,1-4.168.262L261.7,148.1l-3.772-1.488c.619-.353,1.392-.518,1.738-1.2a15.889,15.889,0,0,0,1.861-2.846c.683-.7,1.726-.636,2.605-.919a2.343,2.343,0,0,1,1.712-.032c1.324.312,2.644.719,3.97.972a14.945,14.945,0,0,0,3.941-.921c.649-.4.176.727.273,1.078ZM272.773,131.8c-2.037-1.018-4.024-2.151-6.094-3.093-.61-.249-1.336-.206-1.773-.775-1.065-1.029-2.029-2.2-3.116-3.185l-1.787-.6-1.557,3.841a46.465,46.465,0,0,1-4.9,3.214l-1.747,1.043c-4.04-1.719-8.041-3.527-12.044-5.33.266,1.477.718,3.123-.138,4.5.2.805-.435.586-.974.486l-5.706-.832,3.115-14.155a13.509,13.509,0,0,0,4.746-.634c1.392-.339,2.811-.562,4.183-.992,1.591-1.28,3.21-2.549,4.858-3.74.391-.615.4-1.412.937-1.959a6.144,6.144,0,0,1,1.166-2.1,5.138,5.138,0,0,0,1.148-1.859c.341-.838,1.421-.894,1.805-1.69a4.964,4.964,0,0,0,.279-3.755c-.281-.7-1.114-1.67-.423-2.4.444-.436-.975-.534-.777-1a6.558,6.558,0,0,0,.6-3.244,1.233,1.233,0,0,1-.357-1.264c-.034-.663.271-.92.918-.765a4.232,4.232,0,0,1,3.021.755,42.743,42.743,0,0,0,6.453,2.637c1.128.173,2.312.652,3.426.378.943-.441,1.41-1.477,2.279-2.042,1.156-1.116,2.837-1.457,4.014-2.526,2.1-2.159,3.848-4.627,5.776-6.935,1.86-2.273,3.651-4.608,5.554-6.842l3.2-3.04c1.413,1.285,2.737,2.677,4.222,3.876.007.525-.771.523-1.017.942a18.365,18.365,0,0,0-2.877,2.644c-.23.606-.772,1.1-.847,1.744.265.655,1.078.877,1.637,1.245l3.347,1.42c-1.152,1.7-2.235,3.475-3.43,5.133-.793.232.2.674.417,1.025A25.879,25.879,0,0,1,293.2,94.81q2.31,2.888,4.625,5.773c-2.341,1.442-4.655,2.942-7.025,4.326a1.873,1.873,0,0,1-1.321,1.009,13.419,13.419,0,0,1-3.029,1.153c-1.885.171-2.267,2.13-3.179,3.417-.381.513.4,1.09.555,1.633l2.163,3.966-3.461,8.548Z"
-							transform="translate(-141.228 -73.9)"
-							fill="#fcf5e3"
-							stroke="#000"
-							strokeLinecap="round"
-							strokeLinejoin="round"
-							strokeMiterlimit="10"
-							strokeWidth="1"
-							className={
-								state.districtID === "Sarria-Sant-Gervasi" ? "lit-district" : ""
-							}
-						/>
-						<path
-							id="Les-Corts"
-							d="M277.1,403.2a3.121,3.121,0,0,0-1.072.115,4.248,4.248,0,0,1-2.305.883c-.226.075-.489.083-.7.174a13.06,13.06,0,0,0-1.11,1.361c-.27.483.2.932.375,1.371a3.323,3.323,0,0,1,.272,2.1,9.953,9.953,0,0,0-.418,2.109,1.513,1.513,0,0,1-.262.8,11.489,11.489,0,0,0-1.423,1.391c-.348.275-.487.739-.764,1.069a2.1,2.1,0,0,0-.842.684c-.414.447-.65,1.023-1.019,1.5-1.276,1.205-1.961,2.879-2.916,4.31a2.723,2.723,0,0,0,.1,1.259c.067.428.541.695.483,1.157a4.353,4.353,0,0,0,.252,1.973c-.933-.233-1.706-1.1-2.734-.934.179,1.329-.2,2.666-.078,3.989a.579.579,0,0,1-.123.622c-.18.173-.109.515,0,.718.189.185-.117.439-.083.642-.183.113.01.445-.275.57.037.245.09.438-.069.63-.061.238.22.439.181.648-.307.25-.025.669-.142,1,.017.391-.461.384-.722.516a.956.956,0,0,1-.743.088,12.8,12.8,0,0,1-2.428.219c-.288-.065-.6-.082-.79-.339l-.964-.854c-1.22,1-2.473,1.971-3.673,3a4.628,4.628,0,0,0-.986,1.763,2.991,2.991,0,0,0-.39,1.3,10.157,10.157,0,0,0,.189,2.116c.179,1.522.314,3.05.514,4.568.11.451.049,1.01.464,1.314.51.407.946.9,1.425,1.343.191.613-.466,1.167-.208,1.78.086.4.049.87.364,1.189a4.239,4.239,0,0,1,.791,1.473,6.439,6.439,0,0,1,.507,2.34c-.1.273-.384.538-.222.841.126.152.4.3.388.549a20.754,20.754,0,0,1-.672,2.26,1.337,1.337,0,0,1-.431.772c-.676.766-1.4,1.483-2.143,2.188,4.651,4.652,8.98,9.612,13.6,14.293,3.6.424,7.218.727,10.8,1.315-.553,1.681-2.078,2.737-3.287,3.928a5.566,5.566,0,0,0-1.219,1.467l1.5,1.7c4.547.414,9.1.83,13.611,1.536a62.777,62.777,0,0,1,11.757,2.34,100.889,100.889,0,0,0,15.285,5.85,32.26,32.26,0,0,0,3.424.668,21.624,21.624,0,0,0,2.786.322c1.8.072,3.594.219,5.4.186a8.47,8.47,0,0,1,1.142.242,86.3,86.3,0,0,1,9.3-7.048,43.971,43.971,0,0,1,8.521-5.03l-.1-1.073c-1.667-.828-3.313-1.7-4.96-2.563-2.42-1.255-4.846-2.5-7.257-3.77a28.176,28.176,0,0,0-1.588-2.663c-2.726-4.2-5.52-8.361-8.3-12.528l-1.187-1.593c-4.834-2.246-9.566-4.7-14.362-7.025a9,9,0,0,0-3.059-1.29,21.945,21.945,0,0,1-2.339-.752c.559-1.3,1.381-2.632,1.084-4.108-.148-1.583-1.09-2.9-1.728-4.312-.186-.524-.736-1.8.214-1.632a6.69,6.69,0,0,0,6.059-4.5,4.444,4.444,0,0,0-.43-3.566,9.414,9.414,0,0,1-1.3-3.165c-.32-1.91-.639-3.82-.877-5.742l-.151-1.11a5.276,5.276,0,0,0-1.772-.148,8.97,8.97,0,0,1-2.216.135c-.721-.058-1.424-.252-2.138-.367-.507-.349-1.158-.543-1.485-1.1-.213-.239-.062-.6-.107-.89.016-.331-.073-.688.15-.97.236-.5.532-.973.73-1.483a6.813,6.813,0,0,1,.187-.882.906.906,0,0,0,.352-1.009,1.332,1.332,0,0,1,.208-.865c.436-.1.41-.936.731-1.29.142-.205.188-.627.514-.579a1.237,1.237,0,0,0,.074-.85c-.019-.313-.211-.641,0-.924a4.07,4.07,0,0,0,.257-1.375c.018-.25.119-.522-.067-.731a1.391,1.391,0,0,1-.077-1.183,1.287,1.287,0,0,0,.228-.612.589.589,0,0,0,.035-.588,2.715,2.715,0,0,1-.035-.865c-.083-.206-.089-.431-.173-.623a1.065,1.065,0,0,0-.346-.623,2.341,2.341,0,0,0-.242-1.177c-.236-.221-.129-.537-.069-.865-.246-.507-.726-.869-.905-1.416a21.526,21.526,0,0,0-1.414-3.048c-.54-.279-.966-1.091-1.377-1.238-.348.308-.585.747-.948,1.03-.718.335-1.488.567-2.2.895-.279.234-.487.582-.815.737-.394.2-.834.045-1.244.044a1.955,1.955,0,0,1-1.346-.083,8.309,8.309,0,0,0-2.207-.449c-.519.335-.927.829-1.423,1.19-.643.255-1.274.563-1.933.771a5.438,5.438,0,0,1-1.477.825,3.121,3.121,0,0,1-1.2.319,3.547,3.547,0,0,0-1.293.146,12.074,12.074,0,0,0-1.247.413,5.5,5.5,0,0,1-.775.767Z"
-							transform="translate(-151.337 -247.676)"
-							fill="#fcf5e3"
-							stroke="#000"
-							strokeLinecap="round"
-							strokeLinejoin="round"
-							strokeMiterlimit="10"
-							strokeWidth="1"
-							className={state.districtID === "Les-Corts" ? "lit-district" : ""}
-						/>
-						<g id="g687" transform="translate(268.454 282.505) rotate(90)">
-							<g id="g689" clipPath="url(#clip-path)">
-
-								<path
-									id="path695"
-									d="M7.545,34.054l4.015-3.841-.035-4.949.9-2.769,1.107-3.322-.069-1.246V14.155l3.392.035h7.233l.035-3.668.069-3.6c7.317.151,14.66.236,21.976.208.023-2.365.114-4.73.138-7.095l.969.554L48.347.035,50.978,0h8.029l-.346,1.7-2.319.069-.588,3.08L56.273,8.1,59.11,16.4l2.076,5.918-.692.277.8,2.076,1.523,3.6.554,1.7,1.107,2.63,2.18,4.153,2.423,4.776,4.6,7.095.8.865,3.149-1.142.311.761-3.461,2.18,2.146,4.5,1.419,1.8,2.353,4.984-.173,4.049-1.142,2.8-.623,2.665,9.171,26.856.311.415,3.219,9.552.554,1,.519.8.173.658.554,1.281.519,1.28.761.969,14.293,20.142,2.769,3.876,2.492,3.392,4.637,6.506,2.734,3.911.484.692-.8.484-4.845-6.749-3.184,2.215-.381.208-.312-.069-.035-.277-2.076-3.149.381-.311,1,1.488.934-.692.381.519.484-.346-.381-.519,1.73-1.28-1.211-1.523-4.326-5.78-1.211-.069-.1.069-.311-.035-2.63-3.7-1.8-2.838,1.073-.623L95.726,119.4l-1.315.415-1.869-2.457-.623-1.211-.035-.069-.1-.45-.138-.519.208-1.281-.588-.8-1.384-4.084-1.246-.1-1.211.415-.519-.277-.45-.658-.1-.173h.035l.554-.208-8.687-25.16L74.511,71.846l3.5-6.126L76.83,60.287l-2.526-3.5-.658.485.208.242-.138.1-.311-.415.069-.069.1-.069L73.542,57l.035-.035-.1-.069.588-.45-3.253-4.291L69.147,53.4l2.319,3.115-.346.242L72.088,58l.069.035.831-.485.277.381-.519.346.554.727-.554.484-3.7-4.914-.727.554,3.738,4.88-.311.242-.692-.934-1.107.8-3.461-4.637-.8.588-.9-1.177,4.672-3.53-.8-1.073-1.627,1.281-.45-.588-1.142.969-1.834-2.353,2.7-2.111L62.5,42.5l-.277-.692-.554.311.969,2.215-.173.069-.831-1.973-1.869.761-.554,1.765,3.7,4.741-1.177.969H61.5l-1.765.415-.069-.242,1.592-.381-.623-.865-1.488.519-.1-.311,1.384-.415L59.8,48.52l-1.488.554-.1-.208,1.384-.588-.657-.8-1.384.692-.138-.208,1.315-.727-.623-.831-1.315.8-.173-.242,1.315-.8-.727-.865-.969.865-.208-.208,1-.934,1.246-3.772L58,41.114l-.381-.519-.761.069.554-.519L55.58,37.411l-1.315,2.457-.173-.1,1.35-2.561-.519-.8-1.246-.623-1.834,3.5-.242-.1,1.834-3.565-1.557-.831-1.938,3.7-.242-.1,1.938-3.7-1.315-.761-2.007,3.772-.277-.1,2.007-3.807-.9-.519-2.111,3.876-.208-.069L48.9,33.154l-.277-.173-.346.1-.415-.242L46.029,36.2l-.208-.1,1.8-3.322-1.142.311-1.557,2.8-.277-.138,1.419-2.526-1.142.242-.9,1.7-.311-.1L44.54,33.5l-.865.277-1.142,1.8-.173.381L52.846,41.7l2.111,8.341-.277.381-3.288.761-1.246-.8-1.938-7.545L44.817,41.01l-.623-.381-1.281-.692-.484.1,3.149,12.182.485,2.076.45,1.973.554,2.492.727-.208.173.554,11.351-2.8-.138-.588.692-.242,1.038.623.035.208.934,3.911-.588,1-.692.173-.173-.658L50.389,63.16l-1.938.519h-.208l-.208-.208-3.08.727-.173.035-.8-3.322-2.6,1.28-3.565,2.942-.865.692-4.568,3.807-.242.208-.623.658-1.592,1.834L28.725,73.4l-.8-.069-.727-.1H24.26L22.737,73.2c-3.344-.165-6.514-.078-9.863-.208C8.562,67.7,4.3,62.368,0,57.069l.208-2.423,9.413-7.475C9.265,43.233,7.49,38.119,7.545,34.054Z"
-									transform="translate(0.485 0.519)"
-									fill="none"
-									stroke="#000"
-									strokeLinecap="round"
-									strokeLinejoin="round"
-									strokeMiterlimit="10"
-									strokeWidth="1"
-									className="district"
-								/>
-							</g>
-						</g>
-						<path
-							id="L-Eixample"
-							d="M392.774,606.418c-.43,3.795-.038,7.714-.277,11.49.266,4.826-.853,9.852.969,14.466,5.474,5.637,11.217,11.011,16.612,16.681,3.879,3.845,7.847,7.626,11.559,11.628l.069-.8.1-.727v-2.942l.035-1.523c.165-3.344.078-6.514.208-9.863,5.286-4.313,10.621-8.574,15.92-12.874l2.423.208,7.475,9.413c3.937-.356,9.051-2.131,13.116-2.076l3.842,4.014,4.949-.035,2.769.9,3.322,1.107,1.246-.069h3.772l-.035,3.392,3.7.069,10.175.138,2.042-1.834,2.423-2.526.692-.415,10.659-10.694,1-1-.45-.415-.623-1.315.069-.035-.243-.658-.034-.1h-1.35l-2.423-2.872,1.039-.727,3.737,1.9.415-.1.865-.588.554-.173.9-.138c.091-7.671,0-15.384.138-23.049-.265-3.675.179-7.665-.035-11.351-7.337-.059-14.674-.057-22.011-.035l-.035,3.738-.069,3.634-.035,3.772h-3.7l-3.6-.035h-3.6l-3.945-.035.035-3.7c-7.542-.042-15.354.168-22.876-.035q-15.611-7.824-31.251-15.574l-.45.692-.761.035-.727-.554c-4.211,1.716-7.418,4.378-11.178,6.818C403.746,596.041,398.43,601.394,392.774,606.418Z"
-							transform="translate(-227.105 -348.97)"
-							fill="#fcf5e3"
-							stroke="#000"
-							strokeLinecap="round"
-							strokeLinejoin="round"
-							strokeMiterlimit="10"
-							strokeWidth="1"
-							className={state.districtID === "L-Eixample" ? "lit-district" : ""}
-						/>
-						<path
-							id="Ciutat-Vella"
-							d="M401.226,694.72l3.841,4.014c1.667,0,3.343-.048,5.005-.016,2.016.649,4.026,1.317,6.035,1.989,1.67-.123,3.345-.049,5.018-.069-.052,3.541-.03,7.083-.035,10.625q3.634.025,7.268.1c-.151,7.324-.236,14.65-.208,21.976,2.365.022,4.729.114,7.095.138a5.206,5.206,0,0,0-.527,1.022c.208.49.628.928.533,1.494.054,3.395.018,6.79.028,10.185l-1.7-.346q-.035-1.159-.069-2.319c-1.046-.186-2.092-.427-3.139-.579-1.137.207-2.3.314-3.419.586-4.674,1.592-9.346,3.187-14,4.838-.16-.287-.172-.912-.609-.565-1.907.682-3.737,1.557-5.625,2.283a35.023,35.023,0,0,0-4.2,1.652c-2.92,1.557-5.874,3.05-8.821,4.552-2.373,1.555-4.779,3.069-7.132,4.651-.224.259-.648.49-.747.766l2.228,3.565,1.593-.519.9.737-37.46,15.622,25.474-12.748-.869-3.142-4.984,2.353c-1.367-.072-2.745-.091-4.1-.2-1.016-.383-2-.878-3.037-1.187-.81-.167-1.625-.424-2.432-.536l-26.8,9.152a6.975,6.975,0,0,1-2.1.879c-2.638.9-5.288,1.767-7.92,2.68-.624.315-1.21.7-1.805,1.059a32.262,32.262,0,0,0-3.034,1.179c-.666.415-1.249.972-1.918,1.4-7.889,5.613-15.808,11.185-23.674,16.83-3.922,2.9-7.9,5.711-11.9,8.509l-2.127,1.488c-.074-.291-.646-.687-.338-.9l4.1-2.835.286-5.093c-.217-.226-.329-.8.138-.727l3.149-2.076c.436.31.217.524-.18.712l-1,.673.692.934c-.358.211-.675.429-.244.766.209.271.6-.63.766-.044l1.1,1.492c1.269-1.05,2.611-2.005,3.945-2.97l-1.313-3.449,9.929-6.885-.759-1.56,10.645-7.384c-.076-.331-.341-.726-.3-1.018l1.091-.717c.3.152.542.683.908.483,3.951-1.4,7.921-2.761,11.86-4.183.246-.1.641-.584.8-.43.057.51.342.534.747.305,11.843-4.084,23.677-8.2,35.536-12.231l6.126,3.5c1.824-.411,3.674-.763,5.481-1.211l3.447-2.492c-.232-.193-.4-.906-.709-.465-.28-.171.378-.562.485-.289.219-.235.385.494.617.475l4.244-3.217-1.246-1.661-3.115,2.319c-.168-.538-.472-.182-.757.054-.211.239-.7.391-.736.689.061.286.627.677.311.884-.35.531-.415-.57-.724-.239-.284.166-.6.683-.819.179-.492-.3.062-.491.317-.714l4.347-3.275c-.237-.162-.4-.857-.695-.619l-4.739,3.63c-.47-.324.057-.512.329-.735.521-.184.235-.494-.017-.8-.156-.281-.658-.615-.131-.791l4.353-3.248c-.12-.283-.645-.623-.494-.868l1.083-.828,3.53,4.672,1.073-.8-1.28-1.627c.231-.228.838-.416.358-.721l-.739-.871,2.353-1.834,2.111,2.7c1.682-1.272,3.343-2.581,5.039-3.829.34-.137.856-.211.434-.616-.154-.4-.622.13-.922.164l-1.4.613c-.178-.351.54-.328.75-.518l1.154-.486-.761-1.869-1.765-.554-4.741,3.7a12.973,12.973,0,0,1-.969-1.236,18.961,18.961,0,0,0-.415-1.948c.428-.238.288.495.435.737.127.245.049,1.045.428.612.253-.235.819-.372.508-.785l-.4-1.153c.453-.318.39.356.533.634.1.351.146.881.531.4.349-.205.675-.418.363-.825l-.389-1.044c.371-.267.365.491.555.714.123.413.269.732.606.265.4-.207.5-.477.192-.833-.1-.305-.429-.669-.4-.94.312-.254.382.49.588.672.159.393.338.694.676.253.419-.2.519-.452.175-.795-.1-.3-.581-.68-.424-.928.3-.247.394.487.606.656.184.38.374.724.7.25.282-.229.711-.447.226-.731-.128-.275-.794-.566-.422-.827.258.085.511.52.767.74.31.407.895.387,1.333.595l2.647.874c.044-.394.617-.439.647-.776-.065-.442-.11-.838.266-.3.235.436.586-.146.884-.254l2.045-1.372-2.457-1.315c.169-.375.6.194.89.241l1.775.935c.354-.273.859-.411.981-.89.065-.3.528-.719.333-.93L396,739.018c.085-.5.559.085.841.137l2.827,1.455.831-1.557-3.7-1.938c.086-.5.557.089.839.142l2.968,1.554.761-1.315-3.772-2.007c.06-.538.545.038.837.11l3.073,1.62c.126-.3.478-.666.467-.928l-3.824-2.083c.081-.439.579.164.854.209l3.126,1.66c.339-.275-.1-.651.248-.93.085-.282-.565-.36-.768-.563l-2.525-1.38c.135-.435.595.159.885.216l2.541,1.376a7.021,7.021,0,0,0-.363-1.171l-2.751-1.529c.1-.536.524.014.81.1l1.855,1.042a6.291,6.291,0,0,0-.295-1.17l-1.643-.872c.026-.573.493-.04.783.051.243.074.83.559.842.355-.145-.341-.125-.811-.541-.943a17.726,17.726,0,0,0-1.88-1.125l-5.745,10.486-8.341,2.111c-.394-.159-.465-.533-.528-.912l-.614-2.653a11.371,11.371,0,0,1,.853-1.261l7.487-1.923c.845-1.573,1.693-3.144,2.581-4.692.174-.323.433-.633.239-1.009-.207-.143-.682.149-.992.173-4.064,1.035-8.114,2.129-12.188,3.126-1.85.435-3.7.849-5.56,1.259.049.328.426.833-.132.833-.435.065.016.629,0,.93l2.59,10.489c.385-.146.723-.206.745.309.217.353-.226.638-.341.956-.1.418-.529.368-.865.472l-3.45.824a6.764,6.764,0,0,1-1.018-.646c-.094-.35-.311-.773.242-.74.495-.008.048-.578.062-.874-.921-3.721-1.75-7.466-2.747-11.168.08-.347.289-.591.072-.992-.207-.893-.429-1.782-.625-2.677l3.322-.8c-.475-.9-.861-1.866-1.394-2.733-2.525-3.048-5.011-6.131-7.578-9.144a38.815,38.815,0,0,1-3.234-3.061c-.366-.229-.275-.609-.231-.974a35.114,35.114,0,0,0,.125-4.171c.133-3.363.138-6.729.2-10.093-.29-1.165.959-1.569,1.643-2.25,4.758-3.867,9.546-7.7,14.308-11.562l2.423.208,7.475,9.413C392.521,696.359,396.766,694.725,401.226,694.72Z"
-							transform="translate(-167.346 -404.186)"
-							fill="#fcf5e3"
-							stroke="#000"
-							strokeLinecap="round"
-							strokeLinejoin="round"
-							strokeMiterlimit="10"
-							strokeWidth="1"
-							fillRule="evenodd"
-							className={state.districtID === "Ciutat-Vella" ? "lit-district" : ""}
-						/>
-						<path
-							id="Horta-Guinardo"
-							d="M511.324,347.407c.037.245-.152,1.022-.306.47-.2-.079-.186-.619-.3-.88-.728-2.81-1.394-5.638-2.171-8.435a20.253,20.253,0,0,1-.668-2l-1.987-7.154c-.2-.083-.177-.65-.3-.916.183-.086.64.237.941.171.556-.186.722-.819.89-1.319a13.032,13.032,0,0,0,.218-1.785,26.674,26.674,0,0,1,1.152-2.591c.253-.306.731-.327.939-.7.671-.8,1.2-1.8,2.2-2.233,1.4-.529,2.815-1,4.2-1.563a.947.947,0,0,0,.45-1.038,18.4,18.4,0,0,0,2.27-.191c.56.033.965-.4,1.474-.563a1.994,1.994,0,0,1,1.481-.283c.381-.362.985-.246,1.344-.649.317-.223.86-.124,1.044-.528.632-.17,1.168-.6,1.8-.8a7.8,7.8,0,0,0,1.427.319,5.945,5.945,0,0,0,1.364-.426l9.217-.481c-.09-.389-.04-.681.415-.692,1.179-.268,2.389-.368,3.565-.658.993-.078,2.012-.079,2.99-.205.488-.25.721-.783,1.212-1.037a20.542,20.542,0,0,1,2.279-1.025c.535-.161.9-.686,1.5-.676.363-.07.651-.349,1.032-.4.466.045.526-.263.574-.671a7.741,7.741,0,0,1,1.014-1.982c.463-.559,1.263-.559,1.837-.963a28.047,28.047,0,0,0,2.71-1.6,7.309,7.309,0,0,0,2.108-2.281,8.086,8.086,0,0,1,1.992-1.936,2.6,2.6,0,0,0,2.785-1.378c.216-.635.264-1.5.973-1.81a2.87,2.87,0,0,1,1.281-.565c.386-.1.8-.155,1.024-.525.355-.6,1.012-.55,1.636-.633.807.032,1.559-.883,2.3-.6a3.32,3.32,0,0,0,1.165.691,4.29,4.29,0,0,1,1.366.773,2.974,2.974,0,0,0,1.68.089c.528-.1,1.174.026,1.543-.458a8.147,8.147,0,0,0,1.138-1.153c.895.292,1.869.692,2.765.175a2.526,2.526,0,0,0,1.013-.349c.444-.166.9-.222,1.339-.385a10.228,10.228,0,0,1,1.337-.388c.543,0,.678-.7,1.194-.75a15.411,15.411,0,0,1,1.523,2.526,1.171,1.171,0,0,0,.946.366,3.333,3.333,0,0,0,1.131-.02,9.655,9.655,0,0,0,.692,1.177c.358,0,.716-.024,1.073-.035a34.95,34.95,0,0,0,4.293,5.615c.573.622.343,1.467.275,2.206-.04.4.315.554.594.753a8.059,8.059,0,0,0,2.083.769,5.75,5.75,0,0,0,1.919,1.065c.37.623.406,1.668,1.142,2.152.951.329.882,1.483,1.933,1.494a4.312,4.312,0,0,1,2.156,2.131,4.62,4.62,0,0,0,.149,2.007,1.8,1.8,0,0,0,.437,1.039c.315.2.319.675.5.993.1.5.718.291,1.082.334a2.575,2.575,0,0,1,2.183.8,19.482,19.482,0,0,0,2.374,2.229c.419.246.441.774.79,1.087a3.013,3.013,0,0,0,1.979.627,6.311,6.311,0,0,0,1.7-.523,4.6,4.6,0,0,0,1.71-.827,3.316,3.316,0,0,1,3.583.49,3.256,3.256,0,0,0,1.476.491,5.6,5.6,0,0,1,1.715.54,2.96,2.96,0,0,1,1.222,1.543c.167.47.286.988.461,1.435.682.605,1.66.535,2.5.749l2.162.393c.111.349-.153,1.85-.95,2.126-.4.1-.548.587-.995.537a41.885,41.885,0,0,1-4.513,1.14,7.137,7.137,0,0,1-2.315.012,2.118,2.118,0,0,1-.848-.632c-.4-.176-.516.385-.4.664a1.321,1.321,0,0,1-.045,1.081c-.475.781-1.446.955-2.2,1.369-.494.354-.968-.2-1.4-.4a5.583,5.583,0,0,1-1.7-1.225.4.4,0,0,0-.618.423,3.54,3.54,0,0,1,1.01,2.495,1.437,1.437,0,0,0,.314,1.03,1.8,1.8,0,0,1,.3,1.472c-.087.7-.821,1.127-.963,1.8a1.568,1.568,0,0,1-.129,1.251c-.18.492-.682.8-.8,1.321a5.242,5.242,0,0,0-.691,1.651c-.316.767-.6,2.29.187,2.67.212-.041.886-.2.721.168a1.023,1.023,0,0,0,.115.947c.283,1.034.5,2.085.8,3.115.451.169.375.736.549,1.114a2.106,2.106,0,0,0,.731,1.135c.148.527-.314,1-.768,1.175-.46.233-.7-.391-1.046-.582a1.5,1.5,0,0,0-1.42-.345,3.065,3.065,0,0,0-1.85,1.3,7.1,7.1,0,0,0-1.157,5.291c.135,1.072.347,2.132.5,3.2-.413.092-.119-.739-.461-.714-.529.053-1.056.114-1.581.2.331.58.948,1.8.554,2.215q-.629.15-.674-.5c-.124-.436-.575-.028-.867.005-.338.115-.848.155-.469.579l1.18,2.541c-.306.2-.663.317-.969.519-.7-.381-1.667,1.046-2.5,1.238-.539.57-.484-.9-1.014-.145-1.991,1.265-4.2,2.28-5.885,3.958-1.259,1.391-1.389,3.356-2.193,4.985.454,2.088.613,4.224.934,6.333a3.031,3.031,0,0,1-1.245-.381c-.226.207-.565.115-.854.208-.4.019-.6-.452-1.012-.389-1.4-.186-2.816-.247-4.224-.372-.229.35.111.71.259,1.027a43.781,43.781,0,0,0,3.522,4.856c1.2,1.573,2.273,3.246,3.449,4.841a19,19,0,0,1,2.371,4.406c1.644,3.636,3.225,7.3,4.913,10.917a14.718,14.718,0,0,1-.406,2.316c-1.059,5.219-1.9,10.478-2.792,15.729.384.3.42.828.624,1.246l-.173,1.35c-1.013.138-2.086-.4-3.039.058a35.436,35.436,0,0,0-6.523,5.513c-.559.491-.931,1.258-1.542,1.634a4.879,4.879,0,0,1-1.839.2c-3.381.013-6.758-.185-10.139-.138q-13.151-.122-26.3-.069c-.042-4.639-.088-9.287.026-13.921.288-.6-.208-1.208-.05-1.823a47.206,47.206,0,0,1,1.007-4.87,15.216,15.216,0,0,0,.5-2.444l.7-3.037c.212-.316.635-.224.831-.554a1.357,1.357,0,0,0,.163-.971,4.454,4.454,0,0,0-1.548-.551,5.4,5.4,0,0,1-.776-1.8,17.472,17.472,0,0,1-.02-2.562,13.819,13.819,0,0,1,.784-3.8c.276-.79.681-1.526,1.016-2.291a.647.647,0,0,0-.939-.516c-.946.3-1.864.7-2.833.931-.739-.18-1.374-.729-2.168-.653-.5-.062-1.27.09-1.476-.511-.084-.447.11-.937-.09-1.35.055-.906-.045-1.814.065-2.72a19.19,19.19,0,0,0,.141-3.129c0-.377.132-.792-.211-1.06l-.83-1.1c.47-.585.189-1.387.2-2.073a4.031,4.031,0,0,1,.206-1.489c.289-.588,1.019-.69,1.536-1.007a7.714,7.714,0,0,0,2.979-1.242c.625-.465,1.21-.981,1.8-1.492-.152-.512-.741-.69-.958-1.173-.5-.689-1.031-1.362-1.511-2.059a10.931,10.931,0,0,1-.888-1.117c.1-.391.741-.09.924-.451.586-.63,1.248-1.2,1.775-1.879.12-.554-.48-.8-.778-1.162-.607-.541-1.146-1.174-1.8-1.656-.765-.36-1.646-.447-2.4-.827-1.2-.678-2.39-1.378-3.575-2.075a1.979,1.979,0,0,0-1.524-.222c-.572.057-1.153-.059-1.726.024-.381-.154-.368-.549-.332-.9.088-1.706.064-3.415.09-5.123-.162-.754-1.016-.857-1.605-1.133a7,7,0,0,0-1.541-.584c-.33-.067-.673-.3-.968-.028l-2.046,1.053c-.662-1.514-1.371-3.01-2.027-4.525-.155-.372-.307-.845.067-1.141a.763.763,0,0,1,.963-.324c.552.122,1.2.54,1.7.076.492-.268.568-.892.864-1.329a6.836,6.836,0,0,0,1.118-3,3.167,3.167,0,0,0-1.149-2.42,1.211,1.211,0,0,1-.087-1.5,17.769,17.769,0,0,0,1.177-1.35,2.514,2.514,0,0,0-.265-1.56c-.2-.293-.164-1.055-.677-.79-.3.749-.146,1.865-1.332,1.867a6.212,6.212,0,0,1-2,.027,5.877,5.877,0,0,0-2.45,1.212c-.346.5-.053,1.167-.176,1.731a5.477,5.477,0,0,1-.933,2.663c-.135-.241.033-.67.015-.982.037-.4.135-.815-.157-1.148a2.52,2.52,0,0,0-1.443-.978,5.015,5.015,0,0,1-.706-.931c-.621-.441-1.6.089-2.126-.682-.549-.617-1.354-.875-1.944-1.433-.4-.225-.709.279-.778.609-.29,1.2-.523,2.433-.816,3.623a.736.736,0,0,1-1.054.179,11.287,11.287,0,0,0-2.058-.535c-.432-.265-.308-.8-.083-1.16A1.434,1.434,0,0,0,519,344.93c-.085-.367-.205-.7-.638-.7-.661.378-1.979,1.063-2.249-.269-.179-.446.242-.83.234-1.262-.288-.546-1.15.071-1.392-.535-.121-.436-.6-.366-.946-.419a1.264,1.264,0,0,0-1.129-.112c-.374.2-.313.566-.117.879a2.1,2.1,0,0,1,.022,1.447c-.238.456-.794.393-1.223.425-.475.244-.281.9-.235,1.333A4.368,4.368,0,0,1,511.324,347.407Z"
-							transform="translate(-288.068 -191.602)"
-							fill="#fcf5e3"
-							stroke="#000"
-							strokeLinecap="round"
-							strokeLinejoin="round"
-							strokeMiterlimit="10"
-							strokeWidth="1"
-							className={state.districtID === "Horta-Guinardo" ? "lit-district" : ""}
-						/>
-						<path
-							id="Gracia"
-							d="M538.592,401.958a3.116,3.116,0,0,1-2.138.416c-.3.164-.462.23-.6-.06a7.637,7.637,0,0,0-2.04.816c-.33.2-.748.426-.758.87a6.685,6.685,0,0,1-.488,3.153c-.15.195-.238.959-.522.758.034-.563.12-1.136.122-1.692a2.742,2.742,0,0,0-.706-.921,5.861,5.861,0,0,0-.962-.458,3.071,3.071,0,0,1-.634-.874c-.2-.133-.4-.3-.66-.242-.3-.042-.648.141-.9-.094-.595-.315-1.061-.813-1.649-1.138a5.148,5.148,0,0,1-.911-.671c-.192-.083-.417-.1-.515.126a1.5,1.5,0,0,0-.392,1.084c-.126.738-.32,1.46-.518,2.181a7.818,7.818,0,0,1-.155.823.641.641,0,0,1-.832.282,7.057,7.057,0,0,0-1.6-.469,2.648,2.648,0,0,1-.747-.2.578.578,0,0,1-.2-.705c.116-.444.485-.81.449-1.29a3.846,3.846,0,0,0-.275-.962c-.011-.264-.315-.221-.5-.306-.227-.12-.393.164-.589.246a2.921,2.921,0,0,1-.844.332,1.407,1.407,0,0,1-.668-.515c-.119.11.3,1.032.853.793.448-.189.943.279.438.553-.324.285-.709.514-1,.822-.218.409-.487.8-.673,1.219l-.262,1.178,1,1.454,2.18.658a4.611,4.611,0,0,1,.538,1.38c.37,1.225.772,2.446,1.122,3.673-.092,1.619-.132,3.241-.319,4.853l-.233,2.415-1.661.969-.415,3.911c.156.428.352.849.482,1.281-.072,1.359-.041,2.722-.07,4.083-.023,1.294,0,2.589-.085,3.881-.02.713,0,1.43.043,2.142.027.233.012.521.239.651.956,1.358,1.94,2.7,2.991,3.986a29.005,29.005,0,0,1,2.478,3.188c.554,1.052.054,2.239-.411,3.222a6.973,6.973,0,0,0-.717,1.676c-.429,1.724-.912,3.436-1.328,5.163a9.055,9.055,0,0,0-.324,1.948c-.048,1.013-.077,2.027-.113,3.04h-.9c-.021.217.033.516-.279.415-.273.067.037.232.07.381.031.185-.114.511.036.6.271-.017.063.343.1.508.311.1.2-.261.238-.453.082-.139.413.019.595,0,.13.15.029.428.04.63-.062.781-.06,1.57-.125,2.347a2.214,2.214,0,0,1-.571.919,2.109,2.109,0,0,0-1.07.873c-.544.888-.989,1.841-1.522,2.731-.252.185-.464.449-.8.473a3.79,3.79,0,0,0-1.333.476c-.353.245-.748.455-1.076.722a15.886,15.886,0,0,0-.716,2.262c-.255.977-.595,1.934-.758,2.932a2.288,2.288,0,0,1-.486.92,4.886,4.886,0,0,1-.985,1.026,3.537,3.537,0,0,0-.749.753c-.263,1.766-.579,3.524-.808,5.295-.129.929-.277,1.857-.365,2.79-.465,3.007-.5,6.149-1.742,8.976-.014.329-.7.83-.284,1l7.7,3.852c7.624.2,15.251-.007,22.876.035q-.017,1.851-.035,3.7c2.7.034,5.4.034,8.1.04,2.248.03,4.5.033,6.743.029.1-6.312.162-12.625.1-18.938,0-2.016-.015-4.034.043-6.048a1.413,1.413,0,0,0,.067-.862,2.23,2.23,0,0,1-.086-1.118c.5-3.47,1.576-6.823,2.191-10.271a.863.863,0,0,1,.588-.311.975.975,0,0,0,.412-.8c.038-.223.077-.446-.17-.552a9.135,9.135,0,0,0-1.384-.415,10.844,10.844,0,0,1-.623-1.208,3.558,3.558,0,0,1-.207-1.253,19.983,19.983,0,0,1,.485-4.766,17.641,17.641,0,0,1,.88-2.211,10.775,10.775,0,0,0,.458-1.046c-.054-.245-.183-.507-.466-.522-.361-.091-.7.159-1.047.242-.75.239-1.476.563-2.248.728-.568-.14-1.063-.48-1.627-.625a9.269,9.269,0,0,1-1.658-.152.67.67,0,0,1-.375-.783c-.02-.3.109-.633-.078-.9.081-.773-.052-1.551.042-2.324.034-.982.17-1.959.228-2.936a6.13,6.13,0,0,1-.027-1.384l-1.073-1.419a.981.981,0,0,0,.251-.658,15.133,15.133,0,0,0-.074-1.706,10.413,10.413,0,0,1,.186-1.127,1.486,1.486,0,0,1,.9-.752,4.973,4.973,0,0,1,1.276-.5,10.575,10.575,0,0,0,1.817-.66,28,28,0,0,0,2.353-1.9,1.184,1.184,0,0,0-.48-.611,8.4,8.4,0,0,1-.9-1.136c-.327-.446-.69-.871-.97-1.345l-1.005-1.2c-.043-.3.353-.26.552-.311.3.011.428-.3.633-.462.475-.51,1-.977,1.443-1.515a.53.53,0,0,0-.1-.681c-.763-.721-1.511-1.461-2.283-2.17a8.549,8.549,0,0,0-1.551-.544,3.925,3.925,0,0,1-1.2-.479c-1.1-.628-2.2-1.252-3.276-1.916a2.195,2.195,0,0,0-1.316-.3c-.649.16-1.32-.026-1.977.071-.167-.089-.407-.214-.366-.448,0-.948.082-1.9.067-2.847,0-.918.034-1.842.013-2.757a1.122,1.122,0,0,0-.759-.767,16.026,16.026,0,0,0-3.035-1.1l-2.353,1.211c-.655-1.48-1.328-2.953-1.995-4.427-.084-.339-.35-.724-.111-1.052a1.363,1.363,0,0,1,.652-.574c.573.01,1.095.334,1.66.367a1.747,1.747,0,0,0,.813-.471,16.324,16.324,0,0,0,1.528-3.115,3.4,3.4,0,0,0-.775-3.262,1.285,1.285,0,0,1-.542-.961c-.04-.234.163-.439.224-.66a17.706,17.706,0,0,0,1.177-1.35,2.158,2.158,0,0,0-.124-1.2l-.465-1.191c-.885-.157-.337,1.166-.761,1.523Z"
-							transform="translate(-290.176 -249.752)"
-							fill="#fcf5e3"
-							stroke="#000"
-							strokeLinecap="round"
-							strokeLinejoin="round"
-							strokeMiterlimit="10"
-							strokeWidth="1"
-							className={state.districtID === "Gracia" ? "lit-district" : ""}
-						/>
-						<path
-							id="Nou-Barris"
-							d="M714.775,372.521c-.063.325-.224.669-.005.968.362,1.074.539,2.2.871,3.289.384.12.354.552.478.868a1.992,1.992,0,0,0,.614,1.18c.373.265.214.835-.12,1.063-.23.279-.739.554-1.009.183a2.809,2.809,0,0,0-1.318-.851,3.161,3.161,0,0,0-1.728.565,5.745,5.745,0,0,0-1.577,2.428c-.849,2.225-.094,4.6.2,6.851-.335.046-.215-.485-.341-.7a13.988,13.988,0,0,0-1.7.185c.345.589.9,1.742.582,2.215-.3,0-.658.093-.648-.322-.029-.343-.289-.484-.576-.282l-.984.328,1.35,2.907c-.306.2-.663.317-.969.519-.328-.487-.971.455-1.4.574-.462.174-1.194,1.066-1.565.6-.212-.67-.8.215-1.2.307a30.19,30.19,0,0,0-5.1,3.418c-1.374,1.394-1.5,3.443-2.329,5.134.439,2,.592,4.046.9,6.067.185.448-.339.154-.581.154-.263.035-.533-.408-.731-.166-.26-.013-.494.092-.762.105-.38.026-.566-.427-.946-.382-1.42-.191-2.854-.253-4.281-.379-.214.284.043.555.138.831.682,1.1,1.422,2.166,2.177,3.219a57.068,57.068,0,0,1,3.6,4.8c.831,1.217,1.612,2.474,2.464,3.674a8.866,8.866,0,0,1,.8,1.613c1.785,4.014,3.545,8.04,5.4,12.023l2.041,1.869,2.769-2.942c-.057-.407.376-.555.588-.831a11.3,11.3,0,0,1,1.91-1.762c.466-.578.666.031.914.485,2.649,4.406,6.468,7.906,9.808,11.762l1,.138c.984-3.085,1.755-6.243,2.852-9.289a4.025,4.025,0,0,1,1.012-1.369,9.435,9.435,0,0,1,2.4-1.618c.495-.329,1.169-.151,1.632-.56,8.684-4.981,16.956-10.651,25.713-15.508a3.524,3.524,0,0,1,1.116-.994c2.89-1.779,5.8-3.53,8.676-5.328.86-.569,1.7-1.174,2.575-1.72a6.714,6.714,0,0,0,1.742-.951c1.543-.984,3.088-1.963,4.671-2.881a46.83,46.83,0,0,0,7.735-6.136c1.247-1.236,2.243-2.692,3.368-4.035a18.087,18.087,0,0,0,3.625-7.1c-.043-.576.375-.553.806-.517a5.745,5.745,0,0,1,1.125.484.879.879,0,0,0,.739.387c.245-.029.711.286.785-.008a7.644,7.644,0,0,0,.292-1.532,13.443,13.443,0,0,1,1.278-2.841,16.857,16.857,0,0,1,1.511-2.3c1.67-2.111,3.491-4.1,5.255-6.134,3.78-4.282,7.637-8.5,11.424-12.771a15.113,15.113,0,0,1-1.219-1.055c-.131-.262-.373.485-.545-.052-.355-.388-.765-.746-1.086-1.154a8.418,8.418,0,0,1-.472-2.411,3.477,3.477,0,0,0-1.389-1.641c-.425-.316-.936-.509-1.326-.862-.25-.232-.523-.153-.773-.262-.372-.177-.6-.522-.963-.722-.527-.39-1.2-.519-1.721-.932a7.129,7.129,0,0,1-1.309-.974l-1.794-1.219-.52-2.007A10.628,10.628,0,0,0,798.9,350a5,5,0,0,0-1.022,1.334,9.541,9.541,0,0,0-.421,1.8c.063.309.391.357.666.377a2.426,2.426,0,0,0,1.354.023c.283-.027.6-.239.86-.143.419.136-.182.387-.384.428-.294.042-.511.37-.779.4-.249.152-.434.132-.663.264a1.723,1.723,0,0,0-1.074.184c-.276.837-.51,1.7-.812,2.518-.186.336-.516-.071-.747-.156-.914-.542-1.86-1.043-2.764-1.592-.294-.095-.5-.358-.8-.448-.807-.438-1.574-.962-2.408-1.341-2.212-.814-4.2-2.395-6.623-2.515-1.126.138-2.251.279-3.378.409-1.23-.57-2.626-.459-3.932-.686-.353-.252-.627.2-.926.259-.367-.057-.747.188-1.1.16-.33-.123-.1-.761-.539-.752-.275-.065-.351-.467-.65-.541-.286-.056-.48-.312-.748-.361-.317-.064-.688.046-.929-.224-.753-.1-1.2-.824-1.9-1.064-.8-.343-1.75-.344-2.368-1.043-.248-.227.257-.559.351-.834l1.253-1.9-1-.658-2.353,3.565c.4.224.414.28-.039.127a4.167,4.167,0,0,1-1.312-.473c-.559.174-1.112.41-1.674.547-1.721-.232-3.434-.519-5.143-.824-.964.485-1.9,1.041-2.882,1.482-1.252.434-2.508.857-3.762,1.287a22.413,22.413,0,0,0-.732,3.89,1.744,1.744,0,0,1-.535,1.2c-1.253,1.343-2.541,2.662-3.771,4.02l-1.538,3.417c-1.908.361-3.815.767-5.723,1.1-.441,0-.756-.342-1.1-.58-.59-1.2-1.137-2.425-1.754-3.607-.865-1.172-1.777-2.316-2.613-3.506a4.683,4.683,0,0,0-1.231-1.7c-.19-.375-.505-.115-.5.213a2.192,2.192,0,0,1-.291,1.124c-.243.622-.974.848-1.426,1.216-.8.153-1.557.493-2.361.639a13.256,13.256,0,0,1-3.968.615,1.626,1.626,0,0,1-1.358-.49c-.122-.216-.507-.377-.6-.06-.18.488.293,1.013-.006,1.48a2.446,2.446,0,0,1-1.463,1.122,10.23,10.23,0,0,1-1.128.49,6.8,6.8,0,0,1-1.468-.773,3.71,3.71,0,0,1-1.233-.921.4.4,0,0,0-.684.217c.284.574.968,1.169.95,1.95.129.568-.053,1.237.369,1.7a1.494,1.494,0,0,1,.328,1.228c.071.791-.69,1.261-.948,1.94-.076.336.176.713-.019,1.049a3.938,3.938,0,0,1-.812,1.305c-.174.63-.669,1.131-.763,1.792a3.606,3.606,0,0,0-.207,2.458c.076.326.191.577.548.363A4.921,4.921,0,0,1,714.775,372.521Z"
-							transform="translate(-387.28 -219.311)"
-							fill="#fcf5e3"
-							stroke="#000"
-							strokeLinecap="round"
-							strokeLinejoin="round"
-							strokeMiterlimit="10"
-							strokeWidth="1"
-							className={state.districtID === "Nou-Barris" ? "lit-district" : ""}
-						/>
-						<path
-							id="Sant-Andreu"
-							d="M710.926,469.483c-.214.275-.643.425-.588.831l-2.769,2.942-2.041-1.869c-1.285,5.938-2.24,11.939-3.254,17.927.387.3.418.829.623,1.246l-.173,1.35c-.832.049-1.655-.121-2.482-.138a8.853,8.853,0,0,0-2.4,1.453,38.069,38.069,0,0,0-6.056,5.849c3.08.052,6.161.11,9.24.208.009,6.137-.007,12.275-.139,18.411,2.246-.009,4.491-.05,6.736-.1.263.034.73-.144.612.285q.064,1.692.127,3.383c2.2-1.254,4.369-2.551,6.575-3.785,1.862-.928,3.678-1.95,5.521-2.916,2.19-1.295,4.337-2.664,6.567-3.892.948-.524,1.868-1.1,2.794-1.659.183.2.285.63.622.415a7.185,7.185,0,0,0,1.733-1.182c1.178-.9,2.32-1.851,3.521-2.717l5.475-3.438c0-.277-.2-.639.161-.757a14.866,14.866,0,0,1,3.385-1.726c.925-.374,1.818-.841,2.752-1.182a7.193,7.193,0,0,1,3.629-.15,7.442,7.442,0,0,1,4.885,3.5c1.784,3.09,2.517,6.635,3.3,10.078.414,1.825.728,3.679,1.3,5.464a3.105,3.105,0,0,1,.744-.4c.691-.349,1.015-1.121,1.556-1.645.385-.4.729-.863,1.138-1.216.532-.133,1.106-.2,1.611-.373.281-.22,0-.578-.011-.868l-.365-1.657,1.141-1.107c-.23-.233.185-.506.252-.757a31.676,31.676,0,0,1,2.068-3.054,7.906,7.906,0,0,1,1.141-1.38,18.96,18.96,0,0,0,1.938.727c0,.354.409.329.646.464a8.328,8.328,0,0,1,2.071.936c.543.425,1.054.892,1.613,1.3.889.663,1.872,1.191,2.8,1.8.774-1.72,1.547-3.44,2.249-5.191.412-.238.513-.756.776-1.127.487-.839.93-1.719,1.444-2.533a1.572,1.572,0,0,0,.711-.679q2.847-4.408,5.731-8.791c.667-1.187,1.316-2.382,1.872-3.626a31.006,31.006,0,0,0,1.609-4.083,34.222,34.222,0,0,0,1.041-3.746,124.5,124.5,0,0,0,4.026-31.094,32.832,32.832,0,0,0-.116-3.853c-.136-1.078-.068-2.171-.1-3.255.02-1.048-.062-2.094-.1-3.141.232-.122.029-.55.086-.8-.069-3.309-.179-6.618-.179-9.928q.012-.712.024-1.424c-.172-.063-.171-.595,0-.658a14.463,14.463,0,0,1,.208-2.526c.548.1.182-.973.4-1.335a22.186,22.186,0,0,1,4.83-11.367c-.1-.3-.244-.631-.463-.2-1.254,1.441-2.556,2.842-3.778,4.311a23.526,23.526,0,0,0-2.261,3.169,10.737,10.737,0,0,0-1.09,2.115,5.2,5.2,0,0,0-.5,1.776c-.06.341-.17.674-.25,1.011a10.994,10.994,0,0,1-1.21-.241c-.32-.405-.849-.516-1.282-.762-.257.055-.677-.113-.825.091-.349,1.027-.631,2.081-1.023,3.091a17.91,17.91,0,0,1-1.538,2.822c-.958,1.369-2.02,2.664-3.05,3.979a36.373,36.373,0,0,1-5.307,5.041A43.353,43.353,0,0,1,775,441.043c-1.361.851-2.708,1.736-4.079,2.567a5.464,5.464,0,0,0-1.621.924c-1.98,1.382-4.082,2.575-6.129,3.853-1.47.911-2.963,1.793-4.418,2.721a2.891,2.891,0,0,1-1.112.946c-8.715,4.854-16.93,10.542-25.608,15.457a3.437,3.437,0,0,0-1.439.44,9.94,9.94,0,0,0-2.412,1.7,6.634,6.634,0,0,0-.821,1.247c-1.04,3.036-1.84,6.146-2.8,9.206l-1-.138c-3.544-4.082-7.62-7.793-10.279-12.563a7.982,7.982,0,0,0-1.638,1.341Z"
-							transform="translate(-388.346 -255.191)"
-							fill="#fcf5e3"
-							stroke="#000"
-							strokeLinecap="round"
-							strokeLinejoin="round"
-							strokeMiterlimit="10"
-							strokeWidth="1"
-							className={state.districtID === "Sant-Andreu" ? "lit-district" : ""}
-						/>
-						<path
-							id="Sant-Marti"
-							d="M696.287,667.176c.741-1.986.029-4.108.262-6.155.183-2.067.951-4.034,1.187-6.088-.064-3.511.262-7.074-.476-10.536-.071-.809-.664-1.474-.834-2.2v-2.116c.952-1.911,2.453-3.543,3.115-5.6a6.935,6.935,0,0,0-.378-3.219,2.346,2.346,0,0,1-.082-1.02c.406-.577.65-1.247,1.048-1.829-.192-.086-.017-.46-.21-.662a5.076,5.076,0,0,1-.448-1.069,3.018,3.018,0,0,1-1.79-1.024c-.33-.283-.174-.763-.322-1.135a9.385,9.385,0,0,0-.761-2.1.33.33,0,0,1,.115-.615,1.674,1.674,0,0,1,.958-.32c.558-.231,1.015-.713,1.644-.735a5.887,5.887,0,0,0,1.287-.284,1.862,1.862,0,0,0,1.048-.377,7.6,7.6,0,0,1,.935-.576c-.883-2.9-1.3-5.924-2.09-8.853-.729-2.689-1.5-5.481-3.232-7.721a7.708,7.708,0,0,0-5.6-2.643,8.714,8.714,0,0,0-2.466.413c-1.526.7-3.092,1.322-4.612,2.037a7.335,7.335,0,0,0-1.487.951c0,.275.206.647-.172.74-1.846,1.171-3.715,2.31-5.548,3.5-1.458,1.134-2.914,2.281-4.372,3.408a2.172,2.172,0,0,1-.846.346c-.179-.194-.3-.669-.589-.337-2.182,1.362-4.47,2.543-6.646,3.915-1.284.784-2.562,1.577-3.907,2.253-1.5.8-2.99,1.612-4.5,2.375q-3.079,1.78-6.157,3.561-.071-1.834-.138-3.668c-2.058.006-4.114.1-6.172.093l-1.166.011c.132-6.136.148-12.274.139-18.411-3.094-.091-6.193-.173-9.286-.2a4.263,4.263,0,0,1-1.539.231c-3.533,0-7.066-.141-10.6-.15q-2.127-.019-4.254-.023c.207,3.842-.223,7.69.031,11.531-.136,7.622-.045,15.246-.135,22.869a3.5,3.5,0,0,0-1.925.631,1.294,1.294,0,0,1-.852.351l-3.695-1.882-1.039.727,2.423,2.872h1.35c.053.256.274.568.234.783a5.83,5.83,0,0,0,.379.87,1.863,1.863,0,0,0,.667.873q-5.836,5.844-11.663,11.7a3.437,3.437,0,0,0-1.119.86,42.638,42.638,0,0,1-3.089,3.064,10.588,10.588,0,0,1-1,.851c-4.61-.067-9.221-.112-13.83-.207v7.233q3.634.025,7.268.1c-.151,7.324-.236,14.65-.208,21.976,2.354.018,4.706.108,7.06.138l-.519.969a7.847,7.847,0,0,1,.555,1.168c.06,3.521.023,7.043.033,10.564l1.581-.022q-.013-3.573-.026-7.147l7.222-.05q-.026,2.82-.053,5.64l-5.316-.073q-.046.847-.092,1.695l5.488.07q-.01.7-.019,1.391c-3.095,0-6.19.036-9.284,0l-3-.087q.017.883.035,1.765c1.616.105,3.232.265,4.849.336,1.47-.115,2.949-.181,4.414-.326l6.276-1.706-.174-2.734c-.242-.024-.707.209-.8.009q-.034-2.2-.068-4.4c.181-.375.329-.8.792-.9a6.126,6.126,0,0,1,4.33-.206c1.7,1.026,3.952,1.1,5.3,2.7v2.319c.223-.073.7.094.761-.129-.009-1.217.017-2.439-.013-3.654-.036-.25-.341-.713.123-.564.331-.03.6,0,.474.4q-.016,1.956-.032,3.912c.638.118.746-.246.593-.85,0-1.194-.2-2.6.692-3.553.938-.856,2.3-.78,3.475-.936,3.266-.765,6.638-1.755,10.016-1.15a9.34,9.34,0,0,1,2.391.78,21.808,21.808,0,0,1,.069,2.678v2.617c.229-.044.744.21.62-.2l-.066-4.539h.934a4.35,4.35,0,0,1,0,1.646,13.741,13.741,0,0,0-.143,2.871c-.12.282.625.387.453.021q.016-1.041.031-2.082a1.864,1.864,0,0,0,.527-.955,1.226,1.226,0,0,0-.007-.948,1.794,1.794,0,0,1,1.09-.664c3.548-1.029,7.369-1.84,11.029-.973.268.131.5.289.529.612.127.3.418.549.325.906q-.015,1.011-.03,2.021a.371.371,0,0,0,.658,0c.025-1.12,0-2.252.054-3.365a1.491,1.491,0,0,0,.119-1c3.967-1.229,8.193-3.275,12.446-2.495a1.414,1.414,0,0,1,.777.629c.279.213.081.517.034.792l-.557,2.285c.274.209.512.329.52-.122l.8-2.785c.229-.179.4-.5.734-.415l6.343,0c1.018.021-.279,5.446.734,5.315.147-.907-.312-3.479,1.062-3.508.765.075,7.252,2.95,8.016,2.978.73-.147.593,2.544,1.311,2.368.253-.1-.058-3.687.218-3.454l1.51-.039c.632-.112-.1,2.9.542,2.73.91-.151,5.829-.855,6.758-.821l2.075-1.825-1.4-2.264a15.692,15.692,0,0,0-1.054-3.471c-.1-2.882-.166-5.768-.284-8.646-.4-1.432-.825-2.859-1.2-4.293-.072-.378.439-.3.675-.382.321-.1.731-.035.986-.262.422-.62.051-1.371.04-2.046a8.139,8.139,0,0,1,.615-3.177Z"
-							transform="translate(-330.343 -357.4)"
-							fill="#fcf5e3"
-							stroke="#000"
-							strokeLinecap="round"
-							strokeLinejoin="round"
-							strokeMiterlimit="10"
-							strokeWidth="1"
-							className={state.districtID === "Sant-Marti" ? "lit-district" : ""}
-						/>
-						<path
-							id="Sants-Montjuic"
-							d="M148.381,635.088c-3.371.167-6.718.175-10.092.124l-.138.692-.346.381-.242.969-.462,1.411c-1.535,1.357-2.732,2.945-4.416,3.967a9.812,9.812,0,0,1-2.39,1.267,3.989,3.989,0,0,1-1.834-.519c-.523,1.908-.933,3.852-1.528,5.737a30.969,30.969,0,0,0-1.239,2.99,22.258,22.258,0,0,0-.694,2.52c.693.034.146.742.147,1.152l-.77,3.347c-.439.234-.243.933-.407,1.368a44.6,44.6,0,0,0-.458,4.654,19.551,19.551,0,0,1-3.467-1l-1.378-.455.45-1.661-3.668-1.038-.208,1.557-1.973-.415a1.137,1.137,0,0,0-.727-1.142,11.881,11.881,0,0,1-3.249-.127,18.918,18.918,0,0,0-3.223-.323c.334-.565-.922-.354-.784-.81.337-.4-.8-.684-.878-.4l-1.384-.554c.2-.45-.687-.425-.18-1.068-.028-.366.592-1.173.08-1.208l-3.811-.423-.554,1.7a15.656,15.656,0,0,1-3.621-.977c-11.677-4.526-23.34-9.089-35.026-13.592-2.433-.913-4.9-1.756-7.312-2.713A25.332,25.332,0,0,0,48.9,639c-2.346-.883-4.755-1.6-7.079-2.541A9.908,9.908,0,0,0,39,635.431L37.234,635c3.161,15.409,2.049,29.567-3.046,44.506.4,2.145.745,4.306,1.176,6.44,1.468,3.133,2.874,6.3,4.381,9.411,2.3,4.09,4.629,8.163,7.055,12.179a68.1,68.1,0,0,0,4.044,5.791c1.278,1.652,2.5,3.358,3.812,4.977,4.87,4.428,9.779,8.818,14.625,13.269l15.158,17.7c2.085-.334,4.2-.594,6.265-.974,4.345-2.424,8.662-4.9,13.024-7.291.557.092.546-1.122.939-.539l7.5,7.641,39.983-19.7,1.227,1.76-6.229-9.448.727-2.769a8.2,8.2,0,0,1-2.6-1.63c-.954-.871-1.856-1.833-2.842-2.648l-4.318-1.19-6.541,2.492c-.241-.43-.41-1.052-.977-.608L99.459,725.469,97,719.413c2.556-.856,4.962-2.092,7.444-3.134l1.45-.638-2.284-7.025,5.364-1.938,7.06,6.575,23.395-18.9,1.142,1.281-1.627,1.281q.37.9.98.141c.37-.222.805-.777,1.136-.8.127.419.852.935.252,1.237l-1.122,1.013c.464.427.494.792-.116,1.068l-2.2,1.735c.318.383.905.786.2,1.094l-3.246,2.817a12.107,12.107,0,0,0,3.485,1.193l20.152,3.029c.44-.45.366-1.188.585-1.762l.765-3.152c.525.356.866.262.982-.382.3-.673.414-1.446.739-2.093L163.9,700l1.557.381L161.3,719.344c.419-.01,1.129.441,1.353.121l4.531-18.81,2.769.554c-.149,1.121-.627,2.229-.392,3.367l1.292,12.276-1.523.554c.035.678.772.01,1.168,0,7.158-2.363,14.343-4.648,21.452-7.159l5.413-1.871-.519-9.621,4.984-1.869,3.7,11.005L212,705.6l-2.331-8.109,11.732-3.149,2.608,8.49a21.636,21.636,0,0,0,3.039-.986c.944-.342,1.9-.658,2.844-.987.3-.422-.189-.922-.246-1.379-.234-.32-.137-1.317-.679-1.008-.61.459-.631-.3-.811-.734l-2.278-6.5c.386-.149,1.306-.113.643-.623-.371-.4.567-.638.77-.967l10.319-8.446c.16.626.581.683,1.129.461.461-.12-.124-.885-.111-1.282a1.45,1.45,0,0,1-.017-1.273c-.24-1.033-.495-2.062-.725-3.1l3.322-.8a30.066,30.066,0,0,0-1.437-2.766c-3.4-3.777-6.073-8.254-10.1-11.43-2.94-3.46-6.366-6.461-9.491-9.748q-9.8-9.827-19.632-19.627c-1.828-4.684-.7-9.77-.958-14.643.216-3.768-.158-7.554.266-11.313,3.479-3.191,7.015-6.318,10.486-9.517-1.3-.424-2.683-.15-4.023-.307a37.865,37.865,0,0,1-5.223-.431,29.284,29.284,0,0,1-6.391-1.617c-2.919-.927-5.78-2.023-8.629-3.146a29.2,29.2,0,0,0-6.565-2.58c-3.993-1-8.105-1.325-12.157-1.988-3.46-.456-6.939-.736-10.412-1.074l-1.523-1.73c-.319.512-.132,1.168-.157,1.74.191,1.089.533,2.168.623,3.265-.073.465-.009,1.021-.338,1.241-.224.461-.121.924-.315,1.358-.188.5-.533.988-.317,1.538a22.952,22.952,0,0,1,.467,5.367,8.4,8.4,0,0,1,.15,1.744c.144.585-.089.858-.038,1.224a27.8,27.8,0,0,0,.808,3.517,2.177,2.177,0,0,1,.253,1.32c.309.241.1.881.2,1.3a17.693,17.693,0,0,0-.367,3,16.437,16.437,0,0,1-1.174,4.4c.288.359-.458.784-.515,1.216a32.684,32.684,0,0,1-2.115,3.941,10.546,10.546,0,0,1-2.186,4.691,13.86,13.86,0,0,0-1.977,4.581C148.366,630.835,148.394,632.989,148.381,635.088Z"
-							transform="translate(-34.189 -348.243)"
-							fill="#fcf5e3"
-							stroke="#000"
-							strokeLinecap="round"
-							strokeLinejoin="round"
-							strokeMiterlimit="10"
-							strokeWidth="1"
-							className={state.districtID === "Sants-Montjuic" ? "lit-district" : ""}
-						/>
-					</g>
-					<path
-						id="28-vallcarca"
-						d="M550.777,407.868l-2.353,1.177-1.177-2.63-.8-1.73-.208-.623-.035-.277.138-.277.346-.381.242-.138.381.035.934.277.346.069.346-.1.484-.346.277-.554c1.493-2.346,2.2-4.279.035-6.333l-.1-.588.242-.554,1.038-1.142.138-.208.035-.658-.069-.311-.415-1.523-.934.692-.727,1.28L546,393.3l-1.661.969-.658,2.319-.554,1.592-.277.173v-1.523l-.069-.208-.069-.069-.069-.1-.727-.519-.138-.1-.173-.1c-.8-.313-.871-1.129-1.454-1.246l-.138-.069-.692.035-.1.035h-.1c-.332-.079-.692-.624-1-.831l-.208-.173-.242-.138-.242-.1-.346-.242-.381-.242-.1-.1-.138-.069-.138-.1h-.069a2.548,2.548,0,0,0-.208.865l-.242,1.315-.138.485-.415,1.557-.1.311-.277.1-.208-.035-2.042-.692-.346-.1-.208-.1-.069-.1-.035-.242.035-.1.035-.277.173-.346.208-.623-.1-.623-.208-.208-.069-.415-.588-.173-.623.415-.658.208-.449-.246-.249-.278c.042.249-.074.51.477.721.449.047.925.055,1.075.271l-.093.258-.415.346-.692.519-.692,1.315-.277,1.142c.708,1.052.907,1.5,2.18,1.834l.381.069.208.138.069.1.9,2.319.969,2.284-.069,3.945-.208,3.668-1.834,1.073-.381,3.876.485,1.246-.035.554-.069,6.506-.035.45-.035.865.035,1.557.069.623c1.624,2.312,3.348,4.461,5.157,6.645l.623.934.1.588v.381l-.381,1.557-.277.588-.554,1.073-.381,1.488c-.843,3.023-1.728,6.028-1.523,9.206h-.9l-.035.415h-.346v.208h.1v.035l.035.069v.658l.208.035-.035.485h.208l.035-.519.623.069,3.392.069h.208l1.453.1.173-1.834,1.142-8.133,2.8.519,3.149.623.865-4.672,1.073.831,2.042-1.142,1.35-1.557.381-1.35.658-.969,1.038-1-1-.9-.069-.554v-.554l.138-.519-1.488.727-3.392.069-1.177-.138-5.087-.554-.242-.173c2.924-4.246,2.833-2.463,8.133-1.73l2.976-1.869-1.661-2.007-1.142-.173-1.177-.658.069-2.319,1.177-2.423,2.665-2.042,1.765-.692.519-.658v-.934l-.519-.623h.173l-.035-.311.069-1.73.035-3.738-.1-.277-.208-.277A13.056,13.056,0,0,0,550.777,407.868Z"
-						transform="translate(19713.266 2899.532)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "28-vallcarca"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="29-elcoll"
-						d="M568.681,462.8l-1-.9-.069-.554v-.554l.138-.519-1.488.727-3.392.069-1.177-.138-5.087-.554-.242-.173c2.924-4.246,2.833-2.463,8.133-1.73l2.976-1.869-1.661-2.007-1.142-.173-1.177-.658.069-2.319,1.177-2.423,2.665-2.042,1.765-.692.519-.658V444.7l-.519-.623h.173l.311.242.484-.035,1.107.035.588-.1.415.035.761.311,1.073.658,2.6,1.488,2.042.623.311.173,2.353,2.25.069.173v.311l-.173.242-1.419,1.488-.069.035-.277.277-.623.1-.138.1v.138l1.177,1.488L581,455.529l.415.588.554.485.138.311-.8.692-.761.623-.8.588-1.073.45-1.35.381-1.142.554-.277.242-.173.311-.173,1.038.1,1.557-.069.554-.208.311,1.073,1.419-.831-.277-.588-.208-.554-.277-1.142-.934-1.038-.727-.346-.173-.45-.069-.1.173.035.208.069.208.45,1.038.035.311-.1.415-.415.346-.311-.035-.242-.138-.346-.242-.692-1.073-.519-.761Z"
-						transform="translate(19698.443 2870.975)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "29-elcoll"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="22-vallvidrera"
-						d="M350.909,243.208l.346-.138,1.211-1.038.208-.208.138-.415.138-.692.1-.381.519-.658.415-1.107.623-1.142.485-.346,1.384.554.727,1.384.346.138.658-.761.312-1.073.761-.761.485-.381.277.173.484,1.28.311.1.934-.381.277-.242,1.938-.277.173.242-.346.519-.311.173-.554.658v.692l.277.484-.45.519-.346.45-.1.485-.415.554-.519.623.069.311.588.069,1.177.069.1.311-.381.415-.173.311-.1.415-.554.519-1.35.761-.588.519.035.934.381.865.415.45,1.661.415.727.485.623.45.346-.069.1-.311.035-.588.9-1.661.727-.277,1.211-.381.415-.277.519.242,1.107.415,1.211-.554.865-.485.727.138.484.484.554,1.246,1.177,1.107.554.138,1.177-.588.934-.485L377,247.4l-.069.035-1.315-5.191-.9-3.53-.208-.692-.173-.519-.208-.45-2.25-8.1-.069.035-.173-.727-.069-.277.692.208.208.035.1-.035.346-.208.311-.554.173-.485.173-.969c-.132-1.277.669-2,1.107-3.322l.173-.242.311-.173.242-.1.277-.277,1.107-1.419.484-.485.658-.381,3.6-1.315.588-.242.242-.208.173-.242.069-.277-.035-.311a6.761,6.761,0,0,1-2.526-1.73l-.277-.415-.519-1.523-.831-2.25-.1-.138-.138-.311-.242-.277-.45-.346-.208-.208L376.8,208.7l-.035-.173v-.242l-.138-.277-.415-1.523v-.727l.277-.831.519-1.038.1-.311.069-.138.138-.519v-.242l-.035-.208-.069-.138c-.956-1.088-.955-2.658-1.592-3.911l-.8-.311-.138-.1-.208-1.246-.069-.1-.138-.173-.346-.173h-.173l-.519.1-.277-.069-1.315-1.107-1.9-3.219-.311-.381-1.28-1.73-1.9-1.488-1.142-1.177-1.142-1.834-.311-.311-.173-.242-3.08-1.419-.588-.035-1.177-.381-.277-.173v-1.315l-.1-.346-.277-.242-.069-.311h-.035l-.035-.381-.173-.519.242-.554.277-1.246.035-.415v-.415l.208-.9.1-.208v-.208l-.277-.658-.311-.623-.277-.519-.242-.727-.45-.865-.138-.208-.242-.173-.519-.311.173-.173.1-.346.8-1.765,1.8-1.107.1-.138v-.208l.069-.277-.035-.242-.346-1.073-.346-1.038-.277-.554-.035-.554-.1-.485-.311-.831.1-1.107.069-.1.035-.346-.9-1.142-.865-.623-.208.069-.277-.138-.588-.311-.519-.277-.381.138-.588-.035c-1.376.427-2.7.966-4.084,1.35l-.8.035-.727.069h-.381l-.277-.035-.588.277-.242.035-.519.138-.381-.069-.381.1-.969.588-.346.208-.588-.277-.554-.311-.658-.173-.415.173-.346-.208.035-.035c-.506-.4-.558-.043-1.038.138l.035-.035-1.454.035c-.46-1.477-.124-1.654-.8-3.288l-1.454.45-.415.035-.138.069-.242.069-1.488-.035c-.618-.749-.155-1.742-.035-2.457l-.035-.173-.45-.692-.242-.623-.173-.208-.761-.8-.761-.658-.069-.173.035-.277.208-.381v-.45l-.969-4.222.035-.485.1-.242.346-.208.381-.1.761-.035.208-.069.138-.1.035-.1c.108-1.617-1.466-1.14-2.353-2.353l.1-.346.588-.9.277-.208.761-.242.1-.035.173-.138c.5-.929.151-1.552-.8-1.938-1.054-.393-.992-.121-1.765.208l-.519-.069-.346.173-.415.45-.45.727-.208.1-.242-.1-.484-.346-.346-.692-.173-.519.035-.45.208-.277c.749-.427,1.387-.78,1.627-1.7l-.069-.484-.242-.346-.277-.381-2.007-2.526-.346-.658c-.5-1.87.253-2.723-.623-4.084l-1.384-1.28-.277-.415-.069-.658.242-1,.173-.242.45-.381,2.076-1.142L328,120.8l-.069-.623L327,119.83l-1.938-1-.9-.623.138-.865,4.984-2.976-2.076-.45-2.215-.138-4.015,1.627-1.454,1.281-1.938,3.253-.381.831-.727,2.942-.1.761-.45,1.177-.242.519-.519,1.661-.069.381.242,1.246.346.727.138,2.042.208.658.138.485.277.934.069.485-.069.554-.069.519v.311l-.035.484.069,1.488-.035.035-2.976-1.35-.727-.311-.381-.138-4.188-1.038-.415.346-.934.519-1.177.8-.311.381-.173.069-.381.069-.208.035-1.28,1.107-1,.8c.677,1.005,1.6,3.019,2.526,3.772l1.28.519.242.208.208.346.415.554,1.28.865A8.806,8.806,0,0,1,310.141,148c.2.682.855,1.921,1.177,2.63a5.933,5.933,0,0,0,1.28,1.973,4.681,4.681,0,0,0,.8.865l.1.138.138.138.035.138-.173.208-.484-.208-.519-.311-.381.45-.1.069-.173.1-.415.311-.173.138-.277.173-.173.1c-.509.3-.564.8-1.107.865a14.238,14.238,0,0,1-2.838-1.142l-.277-.173-.346-.208-.242-.138-.208-.138-.173-.1-.138-.069h-.242l-.1-.035-.554-.623-.554-.311-.138-.1-.173-.173-.554-.485-.208-.242-.138-.208a6.024,6.024,0,0,1-1.211-1.869l-.138-.311-.069-.138-.138-.069-.173-.069-.381-.242-.45-.035-.519,1.661v1.592l.346,1.973.138.415.1.8-.035.346-.138.208-.45.208c-4.995,3.833-1.316,2.044-3.634,5.71l1.661,1.488-.173.485-.138.208-.727.381-.45.311.1.831-.069.346-.588,1.107-.381.381-.208.381v.519l-.1.277-.069.1-.242.554.035.692-.035.588-.208.831-.381.623-.173.173-.381.277-.519.277-.415.035-.45-.069-.242-.069-.519-.381-1.211-1.281-.1-.173-.865.45-1.177.934-.1.4-6.956,1.09.381,2.7c-2.028-.05-1.486-.742-3.565.484a2.92,2.92,0,0,0-2.423-3.53c-4.305-1.324-3.785,1.151-6.645-2.7-3.709,1.252-3.72,1.719-5.3,5.191l-.208.138-.588.173-.692.069-.519-.1-.554-.45-1.142-.934-.761-.484-.554-.485-.9-.069-.934,1.523a29.376,29.376,0,0,1-1.107,9.379l.519,2.007,1.557.831.173,1.315v1.419l-.138.934c3.088.294,3.659,2.691,5.087,4.949l.346.242.381.138.761-.1.588.242.45.1h.485l1.73-.173,1.315-.069,1.142-.069,1-.346.485-.346.45-.277,1.419-1.661.242-.1,1,.173.346-.035.346-.138.45.069.346.138,2.25.519.692.1.658.069.173-.415.588-.173,1.834-.242,3.634-.1h.761l.692-.069c1.34,0,2.327.031,3.253,1.107,1.781,1.973.258,2.851,1.211,4.707.8,2.883,2.962,5.008,4.4,7.545,2.947,1.706,3.648,4.771,5.78,7.3l1.488.831,1.384.138.865-.138,1.073-.415.554-.138.381.069.381.138c2.025,1.626,1.818,4.667,3.357,6.576l-.658,1.592-.346,1.523-.069.311-.069.173-.1.138-.831.934-.242.346-.069.519-.173.554-.8,1-1.177,1.211-.346.623-.1.311-.069.311.484.831.381.761.173.588.138.588.242.519-.069.311-.346.865-.692,1.246-.519.554-.658.485-.623.381-1.246.485-.9.277-.761.45-.934.8-.761.45-.761.346c.281.133.185.361.519.485.561.391.925,1.658,1.211,2.215l1.073.519,1.384-.346,1.315-.658.173-.865.311-.208.8.485.9-.1.9-.45.485-.831.588.035,1.073,1.246.934.45.969-.069,1.384-1.142.242-.623.45-.415,1.107.035,1.557-.658,2.561.208,2.457-1.523.242-1-.519-.381-2.111-.173-.761-.692.069-1.592-.8-.346-1,.346-.865-.554-.623-.623-2.423-.208-.1-1.488,1-.865.934-.658v-.934l-.623-.623-.346-1.7.623-.658h.623l1.177-.415.554-.692.623-.969.8.173h.934l.761-.8.727.692.242,1.488.8.623,1.107-.069.8,1.73h1.419l.554-.381.035-3.53.346-.761,1,.311.865-.415.588.346.312,2.146,1.523,1.038,1-.727,1.384-1.627,1.107-.173.969.727.381.588-.519.623-.346.831.658,1.315.519.069,2.561-1.834,1.35-.069.346.346-1.246,2.457-.035,1.315.45.8,1.627-.242.588.588-.415,1.211.242,1.488.658.519.969-.069.692-1.038,1.038-.1,1,.831.8-.1,1.28-.727h1.627l.277.865-.381.969.242,1.038.934.831,1.384.8.519,1.315-1.35,1.523-.242.8Z"
-						transform="translate(19860.496 3048.861)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						className={
-							state.neighborhoodID === "22-vallvidrera"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="path3515"
-						d="M304.7,211.637l-.173,3.98-2.8.277-1.28-.035-8.064,1.869-3.772-1.488,1.315-.658,1.315-1.7.969-1.7.588-.45,2.976-.692,4.637,1.177,2.734-.45,1.661-.623Z"
-						transform="translate(19842.643 2996.473)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						className={
-							state.neighborhoodID === "path3515"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="path3509"
-						d="M292.775,70.226l-.934.761-1,.969L290.283,73l-.381.623-.035.311.1.277.554.415.9.519,3.426,1.454-3.357,5.087-.45.277,1.315,1.315.761.692.8.658,1.523,1.9,3.911,4.88-6.056,3.772-.381.173-.588.381-.173.277-.346.415-1.488.588-1.073.554-1.073.242-1.211.485-.554.277-.415.623-.484.934-.9,1.454,2.907,5.33-3.461,8.548-9.759,7.164-5.157-2.665-.692-.346-.727-.242-.761-.173-.623-.519-2.942-3.08-1.869-.623-1.557,3.841-2.526,1.8-4.118,2.457-4.88-2.111-7.164-3.219.277,1.557.1.761-.035.623-.069.554-.173.519-.242.485-.035.623-6.645-.969,3.115-14.155,1.8-.035,1.454-.173,1.938-.554,1.8-.346,1.938-.519.727-.554,1.869-1.523,1.177-.865.761-.485.277-.242.208-.311.346-1.073.484-.727.519-1.28,1.384-1.73.311-.8.277-.45.277-.242.45-.311.242-.1.242-.173.415-.554.277-.658.138-.969.035-1-.1-.727-.208-.519-.45-.831-.173-.727.035-.242.138-.277.277-.311-1-.658.346-.865.242-1,.138-.831-.035-.623-.069-.208-.311-.311-.035-.173v-.485l.173-.934H257.2l.831.035.934.242.831.554,1.142.554,1.107.519,3.842,1.419.865.242,1.142.138.692.173.45.069.484-.173.554-.277,1.107-1.211,1.107-.934.623-.415,2.353-1.246.346-.242.727-.727,1.073-1.211L287,67.942l3.392-3.219,3.53,3.322.588.45.1.1v.208l-.138.138-.415.208Z"
-						transform="translate(19871.805 3075.278)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						className={
-							state.neighborhoodID === "path3509"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="23-sarria"
-						d="M375.836,430.182l-5.122-2.492-3.149-1.557-3.841-1.9-1.419-.692c-.8-.861-4.433-1.561-5.4-2.042.645-1.559,1.414-2.526,1.038-4.361-.3-2.3-2.214-3.953-1.973-5.676a6.849,6.849,0,0,0,6.437-4.118c1.016-2.926-.577-3.888-1.488-6.437l-.519-3.149-.138-.831-.484-3.565-1.28-.208-2.007.242-1-.069-1.834-.346-1.107-.658-.208-.208-.277-.381v-1.419l.865-1.7.173-.934.277-.277.1-.208.035-.1v-.242l-.035-.208v-.311l.035-.1v-.1l.1-.242.035-.035v-.035l.035-.035h.173a6.254,6.254,0,0,1,.485-1.142l.277-.554.1-.138.069-.035h.138l.138-.485-.1-.588-.035-.242v-.035l-.035-.069v-.1l.035-.1.208-.415.173-1.557v-.069l-.035-.069-.069-.1-.069-.208-.069-.138v-.1c-.151-.709-.042-.313.069-.865l.035-.035.069-.1.035-.035.035-.069v-.1l.035-.138v-.035l.035-.035v-.069l.069-.138v-.1l-.035-.138v-.069l-.035-.035v-.173l-.069-.242V377.2l.035-.069c-.093-.241.023-.05,0-.346l-.035-.069-.035-.138v-.1l-.035-.069v-.069l-.035-.069-.035-.035v-.069l-.069-.208v-.069l-.069-.1-.1-.138-.1-.1V375.5l-.035-.035V375.4l-.035-.035c.122-.175.164-.28,0-.485v-.035l-.035-.069-.035-.138v-.069l-.069-.173-.035-.035v-.069l-.069-.069h-.035l-.035-.035v-.069l-.035-.035v-.138l.035-.069v-.069l.069-.1v-.208a6.827,6.827,0,0,1-1.315-2.423l1.073.519,1.384-.346,1.315-.658.173-.865.311-.208.8.484.9-.1.9-.45.485-.831.588.035,1.073,1.246.934.45.969-.069,1.384-1.142.242-.623.45-.415,1.107.035,1.557-.658,2.561.208,2.457-1.523.242-1-.519-.381-2.111-.173-.761-.692.069-1.592-.8-.346-1,.346-.865-.554-.623-.623-2.423-.208-.1-1.488,1-.865.934-.658v-.934l-.623-.623-.346-1.7.623-.658h.623l1.177-.415.554-.692.623-.969.8.173h.934l.761-.8.727.692.242,1.488.8.623,1.107-.069.8,1.73h1.419l.554-.381.035-3.53.346-.761,1,.312.865-.415.588.346.312,2.146,1.523,1.038,1-.727,1.384-1.627,1.107-.173.969.727.381.588-.519.623-.346.831.658,1.315.519.069,2.561-1.834,1.35-.069.346.346-1.246,2.457-.035,1.315.45.8,1.627-.242.588.588-.415,1.211.242,1.488.658.519.969-.069.692-1.038,1.038-.1,1,.831.8-.1,1.28-.727h1.627l.277.865-.381.969.242,1.038.934.831,1.384.8.519,1.315-1.35,1.523-.242.8.519,1.315-.831,3.392.138,2.042-.1.415-.277.381v.415l-.208.415-.035.346-.035.415.069.45v.45l-.035.311v.242l-.069.208-2.111,1.834-1.038,1.627-.554,1.28-.173.658-2.734.588-.969,4.084-1.765,2.353-.934,2.18-.173,2.25v.1l-.138,1.938-.727,1.142-.934,1.142-.485.934-1.142-.138-4.568-.692-.311,4.4-.138,1.9.208,5.226,1.177,2.077,1.281,1.28-3.253-.173-3.253-.277-1.938.831.346.9-.069.727-.035,3.046V424.4l-.035,5.572v.727l-.415-.138-.831-.381"
-						transform="translate(19808.471 2921.101)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						className={
-							state.neighborhoodID === "23-sarria"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="21-pedralbes"
-						d="M310.493,454.917q-27.724-13.7-55.407-27.479l.277-.485.484-.45,3.461-2.8,1.211,1.073.623.138.934-.035.9-.138.554-.069.277.069.969-.346.138-.1.069-.381-.035-.554v-.1l.173-.208-.1-.242-.1-.208v-.208l.035-.069.1-.1-.035-.138-.035-.208V421.8l.035-.035.173-.208-.035-.277.138-.069v-.242l.138-.242-.138-.208-.069-.277.035-.311.173-.138.035-.173.035-.138v-.1l-.069-.208-.035-.554v-.1l.1-.381.069-.865-.069-.311.069-.623-.035-1.073c1.055-.177,2.083.861,2.734.934l.035-.035-.1-.277-.1-.208-.069-.277v-.969l-.035-.519-.277-.45-.1-.138-.208-.692.035-.865a65.623,65.623,0,0,1,4.188-6.091l.623-.381.311-.519.277-.415.173-.138.761-.831.692-.588.173-.485.1-1.038.381-1.488v-.658l-.035-.45-.138-.588-.277-.485-.346-.761.069-.277.069-.173.761-.934.346-.381c3.516-.435,1.693-1.27,4.949-1.177l.381-.381.381-.415.969-.311.727-.208h.727l.623-.069,1.246-.484.934-.623.346-.1.934-.381.45-.208.208-.069,1.211-1.073.173-.1.831.1.8.138.969.346.415.035.623-.1.45.069h.415l.415-.138.727-.727,1.073-.415.8-.311.381-.173.8-.934.173-.138c.538.694.048.3.727.8.251.438.215.178.554.519a5.862,5.862,0,0,1,.727,1.419l.277.623.277.727.346.727.588.761.1.208v.069c.033.209-.124.22-.1.519l.035.035v.069l.035.035h.035l.069.069v.069l.035.035.069.173v.069l.035.138.035.069v.035l.035.035c.126.266.077.26-.035.45l.035.035v.069l.035.035v.035l.1.1.1.138.069.1v.069l.069.208v.069l.035.035a2.907,2.907,0,0,1,.1.45l.035.069v.138l-.035.069v.035l.035.1-.035.069v.035l.069.242v.173l.035.035v.069c.047.2.066.2-.035.381v.069l-.035.035v.035l-.035.138v.1l-.035.069-.035.035-.069.1-.035.035-.035.173-.1.208v.173l.035.1.035.208v.1l.069.138.069.208.069.1.035.069v.069l-.173,1.557-.208.415-.035.1v.1l.035.069v.035l.035.242.1.588-.138.484h-.138l-.069.035-.1.138-.208.415a7.217,7.217,0,0,0-.554,1.281h-.173l-.035.035v.035l-.035.035-.1.242v.1l-.035.1v.311l.035.208v.242l-.035.1-.1.208-.277.277-.173.934-.865,1.7v1.419l.277.381.208.208,1.107.658,1.834.346,1,.069,2.007-.242,1.28.208.485,3.565.138.831.519,3.149c.911,2.549,2.5,3.512,1.488,6.437a6.849,6.849,0,0,1-6.437,4.118c-.242,1.723,1.673,3.375,1.973,5.676.376,1.834-.393,2.8-1.038,4.361l2.146.692,1.9.623,1,.45.208.138.138.138,1.419.692,3.841,1.9,3.149,1.557,5.122,2.492-6.195,2.561-.138.069-2.976,1.384-3.807,1.211"
-						transform="translate(19860.697 2901.592)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						className={
-							state.neighborhoodID === "21-pedralbes"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="20-lamaternitat"
-						d="M293.944,524.348c-6.145-.925-12.128-1.729-18.342-2.215l-1.523-1.73c1.1-1.746,3.914-3.258,4.534-5.364-3.356-.56-7.361-.911-10.8-1.315-4.548-4.6-9-9.7-13.6-14.293l1.211-1.177,1.211-1.315.761-2.353.069-.415L257.4,494l-.173-.208-.138-.069-.035-.138-.035-.1v-.1l.277-.554-.035-.658-.173-.831-.45-1.281-.173-.381-.658-.934-.1-.381-.035-.415-.069-.242-.035-.173.035-.242.069-.242.208-.588v-.242l-.035-.173-1.211-1.177-.311-.242-.208-.312L253.9,483.1l-.138-1.315-.381-3.461-.173-1.073.069-1,.8-1.869q27.672,13.8,55.407,27.479l-.588,1.35-2.976,2.7-1.627,1.557-1.765,2.665-3.98,6.3-1.938,3.357Z"
-						transform="translate(19861.707 2854.656)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "20-lamaternitat"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="24-trestorres"
-						d="M427.223,482.337l-4.084.658-1.28.692-1.454.8-.346,1.592-.934,4.049-.173.761-.242,1.523-1.7,3.322-2.215,4.43-.588,2.769-2.146-3.322-5.468-8.2-1.211-1.627.415.138v-.727l.035-5.572v-2.734l.035-3.046.069-.727-.346-.9,1.938-.831,3.253.277,3.253.173-1.281-1.281-1.177-2.076-.208-5.226.138-1.9.311-4.4,4.568.692,1.142.138,8.652,1.315,5.849.865Q429.582,473.133,427.223,482.337Z"
-						transform="translate(19779.752 2861.886)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "24-trestorres"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="17-santsbadal"
-						d="M305.611,599.007l-3.115.415-3.392.208-.242-.035.069-1.592h.208l-.035-.588-.035-.9-.138-.035-.069-.519.138-.069-.311-.727-.381-1.107-.208-1.488-.311-1.281.173.035-.069-1.108v-.969l-.138-.069a25.168,25.168,0,0,0-.519-6.229l.519-1.592-.069-.173.277-.969h.173l.138-1.246-.069-.381-.588-2.8v-1.211l.035-.208.138-.242,1.523,1.73c6.214.486,12.2,1.29,18.342,2.215-2.111,3.506-4.192,7.051-6.3,10.555l.1.1C309.512,592.158,307.528,595.559,305.611,599.007Z"
-						transform="translate(19837.996 2800.935)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "17-santsbadal"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="19-lescorts"
-						d="M376.544,559.249l-.692-.173-.45-.069h-1.523l-4.326-.208a37.238,37.238,0,0,1-10.694-2.6,72.923,72.923,0,0,1-10.348-4.222,51.283,51.283,0,0,0-7.025-1.661l2.665-4.568,1.938-3.357,3.98-6.3,1.765-2.665,1.627-1.557,2.976-2.7.588-1.35,3.807-1.211,2.976-1.384.138-.069,6.195-2.561.831.381,1.211,1.627,5.468,8.2,2.146,3.322,1.384,2.076.865,1.557,2.111,1.107,5.953,3.08,1.557.831,2.6,1.315.1,1.073a44.442,44.442,0,0,0-8.652,5.122A84.946,84.946,0,0,0,376.544,559.249Z"
-						transform="translate(19814.166 2828.684)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "19-lescorts"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="18-sants"
-						d="M333.2,614.063l-1.246-1.246-1.142,1.7-1.35-.519-4.637-1.834-.519,1.107-2.18-.865-2.319-.865-1.765-.727-.173-.138-1.073-.865V609.5l-.761-.969-.623-.727-.692-.381-2.111-1.211c2.051-4.346,4.8-8.643,7.129-12.874l-.1-.1c2.107-3.5,4.188-7.05,6.3-10.555l.831.1,2.111.484,1.35.311,1.834.45c3.686,1.511,7.24,3.329,11.075,4.5,3.771,1.282,6.541,2.285,10.59,2.63H354l4.326.208h1.523l.45.069.692.173-3.6,3.288-3.08,2.769-1.073.969-2.734,2.492-.173,2.353-2.6-.138-1.661,1.627-5.26,4.707-2.734-2.146-1.107,1.073-1.488,1.384.588.623-.484.658-.069.208-.069.554-.138.173-.485.415-.208.1Z"
-						transform="translate(19829.717 2796.33)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "18-sants"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="16-labordeta"
-						d="M295.739,649.3l-6.749-.035-5.78-.035.035-.865,2.18-4.637.727-1.073.311-.415.588-1.765.346-1.142,1.7-2.942.45-1.073.1-.208.519-.8-.138-.138a16.941,16.941,0,0,0,1.281-5.537l.242.035,3.392-.208,3.115-.415-1.28,2.6,2.111,1.211.692.381.623.727.761.969v.311l1.073.865.173.138,1.765.727,2.319.865,2.18.865.519-1.107,4.637,1.834,1.35.519,1.142-1.7,2.007,2.007-2.007,1.938-.208.381v1.454l1.35.346-1.107,1.073-1.627,1.488L311.07,649.3Q303.4,649.37,295.739,649.3Z"
-						transform="translate(19845.551 2771.893)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "16-labordeta"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="7-eixampledreta"
-						d="M516.266,642.872l-.45.069-1.523.277-3.565.658-5.4.969-7.475-9.413-.069-2.042h3.807c.118-11.6.2-23.213.311-34.816l9.829,4.914c7.522.2,15.334-.008,22.876.035l-.035,3.7,3.945.035h3.53l-.069,11.455-.035,14.708-3.565-.035H536.2l-.45,1.142-1.107.692-.035,2.353v3.772l-.035,3.634v3.7l-1.246.069-3.322-1.107-2.769-.9-4.949.035-3.841-4.015Z"
-						transform="translate(19729.988 2787.767)"
-						fill="rgba(255,0,0,0)"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "7-eixampledreta"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="13-marinadeport"
-						d="M260.244,711.914l-1.7-.173-6.195-2.457,2.284-6.922-20.176-6.229.1-.381.623-1.315.277-.658.173-.45,1.384-5.364c1.939.888,1.835.657,3.841-.484,2.412-1.408,2.987-2.748,4.88-4.292l.381-1.35.242-.969.346-.381.138-.692c3.374.051,6.735.1,10.106-.069v-5.78l5.78.035,6.749.035,5.953.035v1.9l-3.841,1.9v.831h.554l1.973,1.9,1.211.346,1.142-.657.277.208c-1.86,2.829-1.7,2.511-1.523,5.4l-.831.069a24.523,24.523,0,0,1,.035,4.188l.242.865.242.692.242.8.138.519.208.519.311.415.346.311.1.45.069.831.346,1,.277.8.138.658c.165,1.636.278,4.692-.484,6.126a4.063,4.063,0,0,1-1.384,2.076,1.675,1.675,0,0,1-1.869.346c-1.128-.9-2.1-3.516-3.253-4.187-.95-.779-3.638-1.2-4.811-1.108l-2.111,1.869-.277.554-.761,1.246-.173,1.765.138,1.627-.519.969Z"
-						transform="translate(19871.805 2747.173)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "13-marinadeport"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="8-antigadretadeleixample"
-						d="M473.018,609.493l-.1,11.455h-3.807l-18.342.035-.035-13.393V589.04l-.138-1.627-6.056-8.894,2.6-1.557,2.734-1.315.727.554.761-.035.45-.692c7.14,3.565,14.3,7.066,21.422,10.659C473.114,593.911,473.125,601.71,473.018,609.493Z"
-						transform="translate(19758.668 2800.208)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "8-antigadretadeleixample"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="3-barceloneta"
-						d="M282.994,844.4l.411.846.692-.484,3.911-2.734,6.506-4.637,3.392-2.492,3.876-2.769,20.142-14.293.969-.761,1.28-.519,1.28-.554.658-.173.8-.519,1-.554,9.552-3.219.415-.311,26.856-9.171,2.665.623,2.8,1.142,4.049.173,4.984-2.353.869,3.142-25.474,12.748,37.46-15.622-.9-.737-1.593.519-2.243-3.59.849-.821,7.095-4.6,4.776-2.423,4.153-2.18,2.63-1.107,1.7-.554,3.6-1.523,2.077-.8.277.692,5.918-2.076,8.306-2.838,3.253-.519,3.08.588.069,2.319,1.7.346v-8.029l-.035-2.63-.554-1.073.554-.969c-4.359-.058-8.721-.272-13.082-.069l-8.894-2.215-6.645-1.661-5.26-1.627-.346.069-.208-.208-1.627-1.523.138,3.288,1.246.727.934.623.277.865-1.557-.831-.138.311,1.7.9.277,1.142-2.526-1.419-.138.277,2.8,1.557.277,1.142-3.322-1.8-.069.208,3.357,1.834-.242.415.069.346-.173.277-3.876-2.076-.1.208,3.876,2.111-.519.9-3.772-2.007-.138.277,3.772,2.007-.727,1.315-3.7-1.938-.1.208,3.668,1.938-.831,1.592-3.53-1.834-.138.242,3.53,1.834-.658,1.211-.761.554-2.6-1.35-.069.173,2.457,1.315-2.734,1.8-.519-.519.069.761-.519.381-.138.277L391.8,780.67l-.934-1-.208.208.865.969-.865.727-.831-1.315-.242.173.831,1.315-.831.623-.727-1.315-.208.138.692,1.384-.831.658-.554-1.384-.242.1.588,1.454-.865.658-.45-1.384-.277.1.485,1.453-.831.658-.381-1.592-.277.069.45,1.73-.035.242.969,1.211,4.741-3.7,1.765.554.8,1.869-2.007.831.1.173,2.215-.969.311.554-.692.242-4.984,3.841-2.111-2.7-2.388,1.834,1,1.107-.588.484L386.5,792.3l-1.073.8-3.53-4.672-1.177.9.623.8-4.672,3.461.831,1.107-.934.692.242.277,4.88-3.738.554.727-4.914,3.7.485.554.727-.554.346.519.381-.277-.415-.623-.173-.277h.138l1.246-.969.242.346,3.115-2.319,1.246,1.661-4.291,3.253-.45-.588-.069.1-.035-.035-.069.035-.069-.1-.069-.069-.415.312.1.138.242-.208.485.658-3.5,2.526-5.433,1.177-6.126-3.5L354.04,801.85l-25.16,8.687-.208-.623-.173.138-.633.433c-3.915,1.405-7.885,2.768-11.836,4.167-.366.2-.6-.331-.908-.483l-1.091.717c-.041.292.225.687.3,1.018l-10.645,7.384.759,1.56-9.929,6.886,1.313,3.449c-1.334.965-2.676,1.92-3.945,2.97l-1.1-1.492c-.164-.586-.557.314-.766.044-.432-.338-.114-.555.244-.766l-.692-.934,1-.672c.4-.188.616-.4.18-.712l-3.149,2.076c-.468-.071-.356.5-.138.727l-.286,5.093Z"
-						transform="translate(19845.668 2699.389)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "3-barceloneta"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="1-raval"
-						d="M474.417,724.512l-2.492.554.208.727-.554.173,2.8,11.351.588-.138.242.692-.623,1.038-.208.035-3.911.934-1-.588-.173-.692.658-.173-2.423-10.036-.519-1.938v-.208l.208-.208-.138-.588-.588-2.492-.035-.173,3.322-.8-1.281-2.6-2.942-3.565-.692-.865-3.807-4.568-.208-.242-.658-.623-1.834-1.592-1.073-1.211.069-.8.1-.727v-2.942l.035-1.523.069-2.146.069-2.561-.035-1.073.1-4.084c5.286-4.313,10.621-8.574,15.92-12.874l2.423.208,7.475,9.413-.831,2.492-1.073,3.149-.831,2.423-.346.969-.969,4.984-.138.865-1.073,5.572-.623,3.253.035.692-2.872,6.264-1.246,2.734Z"
-						transform="translate(19751.801 2744.992)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						className={
-							state.neighborhoodID === "1-raval"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="2-gotic"
-						d="M513.716,729.286l-5.745,10.486-8.341,2.111-.381-.277-.761-3.288.8-1.246,7.545-1.938,1.834-3.392.381-.623.692-1.28-.1-.485L497.45,732.5l-2.077.484-1.973.45-.865-3.5,1.246-2.734,2.872-6.264-.035-.692.623-3.253,1.073-5.572.138-.865c.388-4.755,2.7-9.426,4.049-14.016l5.4-.969,3.565-.658,1.523-.277.45-.069C513.192,706.064,513.358,717.792,513.716,729.286Z"
-						transform="translate(19732.816 2736.065)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						className={
-							state.neighborhoodID === "2-gotic"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="4-santpere"
-						d="M566.8,732.971l-5.987.069-8.894-2.215-6.645-1.661-5.26-1.627-.346.069-.208-.208-1.627-1.523c-.163-10.454-.411-20.975-.138-31.424l2.18-.1,3.841,4.015,4.949-.035,2.769.9,3.322,1.107,1.246-.069h3.772l-.035,3.392v7.233l3.668.035,3.6.069C566.856,718.313,566.771,725.656,566.8,732.971Z"
-						transform="translate(19708.561 2736.187)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						className={
-							state.neighborhoodID === "4-santpere"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="26-santgervasi"
-						d="M439.081,537.234l-.1-1.073-2.6-1.315-1.557-.831-5.953-3.08-2.111-1.107-.865-1.557-1.384-2.077.588-2.769,2.215-4.43,1.7-3.322.242-1.523.173-.761.934-4.049.346-1.592,1.454-.8,1.28-.692c6.029-1.184,11.675-1.436,17.719-2.111l4.6.415,1.488.311,1.557.277.381.138,1.177,3.53.069,2.284.484,1.73.831,1.834.484,1.073.277,1.973.9,5.918,1.246.8,1.869,2.492.277.865.969,2.907-.277,1.8-.242,1.454-.588,4.153-.035.484c-.525,3.229-.536,7.261-2.249,10.106-7.774-3.912-15.571-7.75-23.36-11.628l-.45.692-.761.035Z"
-						transform="translate(19769.451 2838.621)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "26-santgervasi"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="27-elputget"
-						d="M518.014,485.19l.035.173-.069,1.142-.069,1.627-.242.554-.312.381-.588.311-.346.311-.346.554-1.315,2.423-.519.415-1.315.346-1.384.9-.242.554-.242.865-.865,3.08-.138.727-.242.623-.8.969-.761.623-.415.45-.969-2.907-.277-.865-1.869-2.492-1.246-.8-.9-5.918-.277-1.973-.485-1.073-.831-1.834-.485-1.73-.069-2.284-1.177-3.53-.277-3.011-.069-.692c1.958-4.352,3.818-8.75,5.745-13.116l.727-.173.727-1.419,1.419.519.311.138,2.423-1.038,1.177-1.454,2.526-3.98,1.419-1.315-.138,7.821.035,1.557.069.623.1.173.069.035c1.639,2.6,3.961,4.91,5.607,7.371l.1.588v.381l-.381,1.557-.277.588-.554,1.073-.381,1.488c-.843,3.023-1.727,6.028-1.523,9.206h-.9l-.035.415h-.346v.208h.1v.035l.035.069v.658l.208.035-.035.485h.208l.035-.519Z"
-						transform="translate(19729.383 2867.097)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "27-elputget"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="31-gracia"
-						d="M538.147,563.777c-6.341-.016-13.034.157-19.346-.035l-7.891-3.945.934-2.076.381-1.523.554-3.5v-.069c.431-3.638.927-7.216,1.523-10.832l.415-.45.761-.623.8-.969.242-.623.138-.727.865-3.08.242-.865.242-.554,1.384-.9,1.315-.346.519-.415,1.315-2.423.346-.554.346-.311.588-.311.311-.381.242-.554.069-1.627.069-1.142-.035-.173,3.391.069h.208l1.453.1.277.035,3.634,1.246,3.5,1.177-.346,2.042,3.841.658,2.8.519,2.492.45,1.8.415-.45,2.561-1.142,6.679-2.146,11.628-5.4-.277-.242,9.275Z"
-						transform="translate(19722.92 2827.513)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "31-gracia"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="30-lasalut"
-						d="M569.257,509.878l-1.8-.415-2.492-.45-2.8-.519-3.841-.658.346-2.042-3.5-1.177-3.634-1.246-.277-.035.173-1.834,1.142-8.133,2.8.519,3.149.623.865-4.672,1.073.831,2.042-1.142,1.35-1.557.381-1.35.658-.969,1.038-1,.692.623.519.761.692,1.073.346.242.242.138.312.035.415-.346.1-.415-.035-.311-.45-1.038-.069-.208-.035-.208.1-.173.45.069.346.173,1.038.727,1.142.934.554.277.588.208.831.277-.035.8.069.658-.173,1.834-.1,1.73.035,1.142-.035.484.1.208-.035.934.035.208.173.242.208.1.346.069,1.315.069.311.1.761.346.519.173.45-.1,2.492-.865.242-.035.346.1.138.173.1.312-.865,1.869L579.6,499.5l-.242,1.35-.173,1.419-.035,1.28v1.073l.242.934.588,1.073.588.138.8.277.208.173-.069.692-.138.242-.242.242-.311.069-.277.242-.727,3.149-4.326-.8-3.357-.588Z"
-						transform="translate(19701.189 2849.123)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "30-lasalut"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="32-elcampdengrassot"
-						d="M580.906,575.351h-3.53l-3.945-.035.035-3.7h-3.53V569.19l.242-9.275,5.4.277,2.146-11.628,1.142-6.679.45-2.561,2.872.588,3.357.588,4.326.8-.138,1.038-.346,1.35-.831,3.6-.208,1.557.208,1-.138.588-.035,2.457c.012,7.526.046,14.965-.1,22.5h-3.7l-3.6-.035Z"
-						transform="translate(19691.131 2819.677)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "32-elcampdengrassot"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="6-sagradafamilia"
-						d="M620.727,628.641l-.415.1c-8.816-4.5-17.859-8.534-26.751-12.874l.069-11.455h.069l3.6.035H601l.035-3.772.069-3.634.035-3.738c7.336-.016,14.674-.035,22.011.035.214,3.686-.23,7.677.035,11.351-.14,7.665-.048,15.378-.138,23.049l-.9.138-.554.173Z"
-						transform="translate(19678.408 2790.612)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "6-sagradafamilia"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="67-vilaolimpica"
-						d="M607.784,770.955l-3.28-.1.035,1.765,4.707.346,4.464-.311,6.368-1.73-.174-2.734-.8.1-.07-4.5.312-.623.553-.277.174-.1.934-.277,1-.1h.934l1.211.277c1.71,1.034,3.945,1.1,5.3,2.7v2.319l.761-.035v-3.7l-.174-.623h.347l.138-12.874-3.841.45.208-12.182-25.887-.519-.138,15.124v1.315l-.035,1.315.035.623c2.365.019,4.73.112,7.095.138l-.554.969.554,1.073.035,2.63v8.029l1.581-.022-.026-7.147,7.222-.05q-.027,2.82-.053,5.64l-5.316-.073q-.046.847-.092,1.695l5.488.07q-.01.7-.019,1.391C613.689,770.939,610.879,770.989,607.784,770.955Z"
-						transform="translate(19674.49 2712.557)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						className={
-							state.neighborhoodID === "67-vilaolimpica"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="68-poblenou"
-						d="M679.287,737.112h-.554l.069,4.776-.623-.035v-4.845l-.069-.45c-3.978-1.935-8.553-.531-12.6.415-4.234.151-4.086,1.612-3.945,5.26l-.623.035.035-4.292-.415-.035.139-12.874-3.842.45.208-12.182.1-7.095,14.6-3.392.173-11.8,22.01,11.213v1.765l.174,18.135-7.268-.415.035.381-.069.069-3.184.138.243,5.226-.589.138-.034,2.63-.554.138.8,4.5-4.326,1.627Z"
-						transform="translate(19644.316 2737.964)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "68-poblenou"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="71-provençalsdelpoblenou"
-						d="M711.571,693.133l-22.01-11.213.1-7.822q27.635.11,55.269.208v4.707l-11.178,2.734v15.089l-11.144.1-2.906.035L715,694.864Z"
-						transform="translate(19626.707 2747.092)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "71-provençalsdelpoblenou"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="65-elclot"
-						d="M672.376,644.209l-7.2-.035-.485.035h-.8l-2.457.035-.345-.588-1.073-1.384-1.592-.588-.969-.311.035-4.811v-3.668h3.772l7.683.035,7.372-7.406.242-.242,4.671-4.707,1.938-1.973.035,2.146-.035,1.142-.069,3.634,3.773-.035,2.63-.069h.934c.329,6.282.058,12.576.242,18.861l-9.759-.035H678.26Z"
-						transform="translate(19643.994 2776.982)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "65-elclot"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="64-campdelarpa"
-						d="M683.214,604.946l-1.938,1.973-4.672,4.707-.242.242-7.372,7.406-7.683-.035h-3.772c.029-8.643.117-17.282.069-25.921,4.449.007,8.812.067,13.254.173h2.216l.969-.242,1.972.035,3.6.069,3.669.1C683.261,597.285,683.335,601.119,683.214,604.946Z"
-						transform="translate(19643.953 2790.635)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "64-campdelarpa"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="72-santmartideprovençals"
-						d="M755.122,642.322c-8.594-.071-17.189-.07-25.783-.1l-.069-7.683V630.9l-.035-3.876.242-.138,6.161-3.565.657-.311,5.191-2.769,4.222-2.561,3.219-1.834,1.765-1.073.345.484.589.969,3.288,5.33C755.015,628.476,755.052,635.4,755.122,642.322Z"
-						transform="translate(19605.34 2779.042)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "72-santmartideprovençals"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="73-lavernedailapau"
-						d="M791.747,629.621H787.87l-3.668-.035h-3.633c-.07-6.92-.107-13.846-.208-20.765l-3.288-5.33-.589-.969.416-.1.554-.312.1-.1,4.257-3.322,5.676-3.565-.069-.588.9-.657.968-.484,2.492-1.073,1.868-.831,1.108-.242a7.36,7.36,0,0,1,5.433,1.281c4.332,2.627,5.4,13.971,6.713,17.823l-.242.1-.38.277-.554.346-.172.1-.1.035-.624.1-.069.035-.138.069-1.454.242-.9.519-.276.138-.52.069-.726.415.034.312.139.138.1.277.1.208.415.969.208.969.069.588.484.588.138.1.52.277.831.277.035.173.242.554.277.554v.381h.1v.069l-.242.381-.485.934-.347.554v.588l.069.035c.009.958.632,1.768.485,3.08l-.1.658-.069.312-.1.277-.347.588-.276.415-.762,1.246-.761,1.35H800.4l-4.811-.035Z"
-						transform="translate(19579.893 2791.778)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "73-lavernedailapau"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="61-lasagrera"
-						d="M750.877,578.34l-.345-.485-1.765,1.073-3.219,1.834-4.222,2.561-5.191,2.769-.657.312-6.161,3.565-.242.138-.035-.9-.1-2.769,3.841-3.149.1-7.718.069-7.268.1-8.306-.416-1.038,1.35-1.35c3.871-1.751,6.665-6.158,10.279-8.79l1,.138.831-2.734,6.126,6.437-.069,2.146,1.834,3.115.934,1.454,1.592,1.419,2.6,1.8,1.453.8,1.142,6.333.1.623.069.588-5.676,3.565-4.257,3.322-.1.1-.554.311Z"
-						transform="translate(19605.5 2815.961)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "61-lasagrera"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="63-navas"
-						d="M709.749,601.3h-.934l-2.63.069-3.773.035c.132-6.127.148-12.28.139-18.411l-3.669-.1-3.6-.069-1.972-.035a38.107,38.107,0,0,1,6.056-5.849l1.592-1.107.761-.346.831.069.692.035,1,.035h7.475l1.73-1.8.416,1.038-.1,8.306-.069,7.268-.1,7.718Z"
-						transform="translate(19624.688 2801.097)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "63-navas"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="59-bonpastor"
-						d="M854.952,526.809c-1.008,4.817-1.237,9.616-3.08,14.258l-.312.865-.692,1.592-.97,2.042-1.142,2.076-1.764,2.665-2.631,4.049-1.523,2.353-.519.346-1.939,3.461-.242.173-.658,1.592-.415.969-1.177,2.63-2.492-1.557-1.281-1-.415-.381-.692-.484-1.281-.485-.9-.346-.069-.242-1.211-.415-.727-.311-.346.311-1.107,1.488-1.315,1.869-.8,1.384.1.138-1.142,1.107.518,2.353-.173.208-1.627.346-1.315,1.453-.658.831-.242.277-.312.242-.519.208-.242.208-.1.035c-1.173-3.587-2.416-14.119-5.848-17.1a7.127,7.127,0,0,0-6.3-2.007l-1.108.242-2.041.865-1.454.554-1.834.969-.9.658-.1-.623,2.906-2.076,1.038-.831,1.835-1.454.865-.831,1.246-.934,1-1.246.52-.9.761-1.454,1.384-3.288,1-3.772.554-1.592.761-.934.484-1.315.622-.969,1-1.384.174-.208.83-.831.554-.519.934-.8.312-.242.726-.554.658-.484c4.463-2.62,8.455-5.407,12.84-8.064l7.2,5.745,4.015-5.33,6.16,1.384,4.915.969Z"
-						transform="translate(19567.25 2829.007)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "59-bonpastor"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="60-santandreu"
-						d="M791.919,520.737l-.761,1.454-.52.9-1,1.246-1.246.934-.865.831c-1.826,1.584-3.841,2.923-5.78,4.361l-1.142-6.333-1.454-.8-2.6-1.8-1.592-1.419-.934-1.454-1.834-3.115.069-2.146-6.126-6.437,1.28-4.5.242-.554.139-.346.311-1.073.658-1.107,1.176-1,.727-.484,1.176-.623.934-.173c8.9-5.078,17.16-10.768,26.13-15.747l.622-.692,8.929-5.468,2.769-1.869,3.945,3.357,1.9,2.457,2.6,2.284,1.557.761,2.008.45.969.727,1.661,1.523.1,1.142,2.733,1.038c-7.065,4.673-14.507,8.835-21.526,13.566-2.18,1.509-4.6,2.571-6.644,4.257a23.223,23.223,0,0,0-2.63,2.388l-.174.208-1,1.384-.622.969-.484,1.315-.761.934-.554,1.592-1,3.772Z"
-						transform="translate(19585.467 2855.222)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "60-santandreu"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="58-barodeviver"
-						d="M871.511,511.82l2.733-1.7,5.849-3.6,5.364-3.426,9.345,3.668-.589,3.149.277.208-.692,4.464-4.915-.969-6.16-1.384-4.015,5.33Z"
-						transform="translate(19528.715 2839.186)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "58-barodeviver"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="57-trinitatvella"
-						d="M881.337,448.246l-2.733-1.038-.1-1.142-1.661-1.523-.969-.727-2.008-.45-1.557-.761-2.6-2.284-1.9-2.457-3.945-3.357.484-.138,4.154-2.63,1.523-.9.588-.346a46.961,46.961,0,0,0,6.437-5.018l1.316-1.281.208-.173,1.072-1.315,1-1.246a18.956,18.956,0,0,0,4.534-8.583h.8l1,.485.312.311,1.177.208a11.8,11.8,0,0,1,1.557-4.43l.865-1.454,1.039-1.384.934-1.177,2.111-2.388,1.592-1.8h.035l.208.45a22.761,22.761,0,0,0-4.984,12.736l-.242-.035-.138.8-.069,1.73-.1.035v.588l.1.035-.035,2.042.069,4.153.138,5.918-.1.035c.165,3.013.082,5.893.346,8.929l-.035,1.938-.1,2.907-.069,1.869-.865,6.506-.035.035Z"
-						transform="translate(19532.836 2894.039)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "57-trinitatvella"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="62-elcongresielsindians"
-						d="M725.69,547.327l-1.73,1.8h-7.475l.173-1.35-.381-1-.242-.242c1-5.892,1.992-12.112,3.254-17.927l2.041,1.869,2.769-2.942.035-.312.381-.346.172-.173,1.281-1.315,1.073-.761c2.819,4.968,6.522,8.233,10.279,12.563-3.613,2.633-6.407,7.04-10.279,8.79Z"
-						transform="translate(19612.449 2827.594)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "62-elcongresielsindians"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="35-elguinardo"
-						d="M650.859,550.987l-3.6-.035V539.878l.069-4.88.761-1.073.865-.415-1.627-2.146-.415-.415-1.9-.554.761-4.707,3.219.554.38-1.9-.588-.138-.415-1.73.484-.069.831.934.035-.381-.623-1.177-1.038-.761-.208-.415.311-1.211v-.519l-.415-.658-1.315-.8,1.141.173,1.143.484.761,1.384,1.108.346.588.035,1.039-.484.865-.623.1-.415-.07-.277,7.718,3.53,1.316-.312,1.765-.311,2.041-.519,1.246-.173,1.972.934L671,519.113l3.565-1.177,2.6-3.288.035.069,3.219,7.2.624,1.315c-1.262,5.815-2.258,12.035-3.254,17.927l.242.242.381,1-.173,1.35c-3.141.047-1.852-.854-4.88,1.315a38.1,38.1,0,0,0-6.056,5.849l-.969.242h-2.215l-1.73-.035h-.518l-3.773-.1h-3.322Z"
-						transform="translate(19650.699 2832.966)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "35-elguinardo"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="33-baixguinardo"
-						d="M628.316,562.5c-6.142.005-12.268-.037-18.411,0l-.035-2.7v-8.652l.035-2.457.138-.588-.208-1,.208-1.557.831-3.6.346-1.35.138-1.038.727-3.149.277-.242.312-.069,3.911,1.177,2.906,1.384,3.98,3.011,2.6.277,1.9.554.415.415,1.627,2.146-.865.415-.761,1.073-.069,4.88Z"
-						transform="translate(19669.643 2821.414)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "33-baixguinardo"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="34-canbaro"
-						d="M620.506,522.963l-2.6-.277-3.98-3.011-2.907-1.384-3.911-1.177.242-.242.138-.242.069-.692-.208-.173-.8-.277-.588-.138-.588-1.073-.242-.934v-1.073l.035-1.28.173-1.419.242-1.35.519-1.454.865-1.869-.1-.311-.138-.173-.346-.1-.242.035-2.492.865-.45.1-.519-.173-.761-.346-.312-.1c-1.609-.171-2.315.293-2.042-1.627l1.107-.381,1.281.311,1.038.346,1.246.069,2.111-.831.727-1.592.346-.138,1.281.277.554.346.208,3.08-.138,1.557-.346.588.208.45.623.138.312-.519-.277-.9.311-.035,3.115-.277.692.761.485,2.007.451.381,1.142.208,1.384-.242.484-.346-.207-1.038-.9-.346L617.08,505l1.316.173,1,.208.311.173.52.173.311.138.416.035.311-.035.865.173.831-.208.553.761,1.766,1.384.692.415.381.381,2.249,1.834.07.277-.1.415-.865.623-1.039.484-.588-.035-1.108-.346-.761-1.384-1.143-.484-1.141-.173,1.315.8.415.658v.519l-.312,1.211.208.415,1.038.761.623,1.177-.035.381-.831-.934-.485.069.415,1.73.588.138-.38,1.9-3.219-.554Z"
-						transform="translate(19675.203 2840.398)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "34-canbaro"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="36-lafontdenfargues"
-						d="M649.607,491.971l-1.593-2.907,1.489-1.627.9-1.661-.138-.485-.416-.242,1.108-2.111-5.71-3.219-2.18.45-.381-1.661,2.457-3.5,1.938.831,1.419-2.734,2.63-3.461,2.353.311,1.765.658,5.71,1.73,3.807.934.347-.173c1.02,2.223,3.186,4.631,4.6,6.576l.242.311.174.208,1.28,1.9,2.112,3.184.208.208.9,2.042,1.28,2.872-2.6,3.288-3.565,1.177-1.834,2.076-1.972-.934-1.246.173-2.041.519-1.765.311-1.316.311-7.718-3.53Z"
-						transform="translate(19651.951 2857.201)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "36-lafontdenfargues"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="37-elcarmel"
-						d="M624.259,476.109l-1.766-1.384-.553-.761-.831.208-.865-.173-.311.035-.416-.035a8.948,8.948,0,0,0-3.461-.865l-.276,1.938.9.346.207,1.038-.485.346-1.384.242-1.142-.208-.45-.381-.484-2.007-.692-.761-3.115.277-.312.035.277.9-.311.519-.623-.138-.208-.45.346-.588.138-1.557-.208-3.08-.554-.346-1.281-.277-.346.138-.727,1.592-2.111.831-1.246-.069-1.038-.346-1.28-.311-1.108.381-.1-.208.035-.484-.035-1.142.1-1.73.173-1.834-.069-.658.035-.8-1.073-1.419.208-.311.069-.554-.1-1.557.173-1.038.173-.311.277-.242,1.142-.554,1.35-.381,1.073-.45.8-.588.761-.623.8-.692-.138-.311-.554-.485-.415-.588-1.073-1.419-1.177-1.488v-.138l.138-.1.623-.1.277-.277.069-.035,2.111-.173,2.146-.381,1.834-.831,3.633-2.976,2.423-.45,2.458.588-1.177-1.973,1.73-1.384-1.418-1.419,2.388.727,3.6.554,4.88.381,5.295,1.938,1.938.658-1.592,1.488-5.468,7.164-2.63,3.461-1.419,2.734-1.938-.831-2.457,3.5.381,1.661,2.18-.45,5.71,3.219-1.108,2.111.416.242.138.485-.52.969-.38.692L623.739,474l1.593,2.907-.381-.381Z"
-						transform="translate(19676.227 2872.267)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "37-elcarmel"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="38-lateixonera"
-						d="M590.644,436.817l1.419-1.488.173-.242v-.311l-.069-.173-2.353-2.25-.311-.173-2.042-.623-2.6-1.488-1.073-.658-.761-.311-.415-.035-.588.1-1.107-.035-.485.035-.311-.242-.035-.311.069-1.73.035-3.738-.1-.277-.208-.277a13.058,13.058,0,0,0-3.5-1.315l2.319-1.142,6.818-3.253,1.973-.934,1.7,3.426,2.353.9-.242,1.661,2.388,3.011,1.73.554,4.361.554c4.274,2.573,2.412-.765,6.021,3.184l-1.73,1.384,1.177,1.973-2.458-.588-2.423.45-3.633,2.976-1.834.831-2.146.381Z"
-						transform="translate(19687.658 2886.122)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "38-lateixonera"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="40-montbau"
-						d="M623.107,329.856l-3.565.1-5.156,1.869-6.472,3.08-6.022,2.907-2.457,1.177-11.455,5.5-.1-.173-2.492-4.776,1.246-3.288,1.142-.623-.173-.831,2.284-1.627-1.765-3.634,4.949-2.146-1.454-11.836,1.073-1.973,1.869-.381,1.142-.8-.658-1.177-2.872,1.523-1.073,1.107-1.246,1.973-1.246-.069-1.28-2.872-2.526-.658-1.281-2.6-1.661-1.038-1.765-2.457-1.35-1.107-1.419-.381-.035-.554.45-.138.623-.138,1.592-.242,1.35-.277,2.942-.173.415-.277.381-.45.658-.485.969-.415,1.488-.623.485-.346.242-.1.588-.1.346-.208.658-.208h.277l.173-.8.242-.554.554-1,.138-.208.208-.208.208-.173,1-.381,1.177-.623,1.765-1.038.969-.761.8-.831.727-1.107.554-.658,1.281-1.038c1.368,0,2.362-.205,2.872-1.592l.139-.658.207-.45.243-.311,1.107-.658.416-.069.761-.208.8-.831a7.288,7.288,0,0,0,3.183-.865l.174.069.172.1.519.415.52.208.934.415.484.381.623.173H614l1.834-.208.242-.138,1-.934.312-.311.035-.1c2.96,1,1.574.07,4.6-.381l.346-.138.45-.1.38-.173.934-.208.243-.173.312-.346.138-.1.242-.069.208.346a20.537,20.537,0,0,1-4.914,3.115l-.554.346-.138.415.1.554.312.519v.727l-.208.969-.519.346-.692.242-.035.623.207.9.138,1.107-.449.761-.069.727,1.142,1,.934,1.177.554.865-.208.381.519,1.35,1.454,1.765,1.073,2.8.623,1.315v1.142l.242.969-1.038,1.28-.658,2.18-.242.831-.347,1v2.907l.589,2.146-.173,2.388-.243,1.869.8,2.215,1.766.173-.07,1.834Z"
-						transform="translate(19687.174 2957.575)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "40-montbau"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="41-valldhebron"
-						d="M614.823,409.9l-2.077-.934-4.361-.554-1.73-.554-2.388-3.011.242-1.661-2.353-.9-1.7-3.426,11.455-5.5,2.457-1.177,6.022-2.907,6.472-3.08,5.157-1.869,3.565-.1,6.056,2.215c2.02,1.575,2.288,1.686,1.764,4.43l-2.388-.519-.276,2.319-1.765.588.9,1.627-2.907,1.177-.416.208-.484-.069-.451.208-.242.415-.1.519-.312.277.243,2.284-.069,1.211-1.108,2.146-1.072-.208-2.007-.035-4.914-.554-2.908,2.872-4.637,1.9-2.111.658-1.039.242Z"
-						transform="translate(19674.691 2903.208)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "41-valldhebron"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="43-horta"
-						d="M674.967,373.139l-.347.173-3.807-.934-5.71-1.73-1.765-.658-2.353-.311,5.468-7.164,1.592-1.488-1.938-.658a9.941,9.941,0,0,0,.311-4.326c.757-3.2-1.317-5.422,1.315-8.894l.069-1.211-.243-2.284.312-.277.1-.519.242-.415.451-.208.484.069.416-.208,2.907-1.177-.9-1.627,1.765-.588.276-2.319L676,336.9c.524-2.744.256-2.855-1.764-4.43l-6.056-2.215,2.076-3.253.07-1.869-1.766-.173-.8-2.215.243-1.869.173-2.388-.138-.519-.451-1.627v-2.907l.589-1.834.658-2.18,1.038-1.28-.242-.969v-1.142l-.623-1.315-1.073-2.8-1.454-1.765-.519-1.35.208-.381-.554-.865-.934-1.177-1.142-1,.069-.727.449-.761-.138-1.107-.207-.9.035-.623.692-.242.519-.346.208-.969v-.727l-.311-.519-.1-.554.138-.415a26.32,26.32,0,0,0,5.468-3.461l.9,1.315.138.277.276.588.415.311.243.035,1,.069.415-.069.208.415.484.761h.139l.934-.035a34.993,34.993,0,0,0,3.876,5.157l.207.208.243.311.173.242.1.277.069.8-.1.554v.277l.069.381.242.173.589.415.831.277.9.242.761.554.345.208.969.381c.505,2.824,1.994,2.026,2.112,3.357a3.938,3.938,0,0,1,2.561,1.627l.208.277.242.415-.035,1.454.035.277.138.277.069.415.277.623.242.138.45,1.142.242.1h1.073l.554.069.449.1.381.173.277.242,1.211,1.28,1.626,1.384.138.242.069.208.347.45.311.242.589.208.761.138h.588l1.766-.658.934-.346.415-.346.346-.173a3.318,3.318,0,0,1,2.8.381c1.708,1.42,2.756.055,4.534,2.146l.207.381.451,1.419.069.311.415.311.485.173,3.806.692.1-.069-.1.311-.069.692-.311.658-.416.45-.38.208-.1.035-.174.208-.172.138h-.208l-1.592.485-2.8.658-1,.138-1.073-.035-.416-.035-.484-.242-.416-.415-.242-.035-.138.138-.069.277.173.934-.173.45-.174.242-.449.381-1.869.865-.312-.069-2.042-1.211-.451-.484-.277-.069-.172.035-.138.173-.069.242c1.1,1.091.956,1.779,1.107,3.184l.415.623.139.311.035.173-.035.831-.139.381-.622.865-.208.415.1.658-.07.277-.242.588-.588.761-.069.242-.554,1.038-.208.761-.242.865-.07.519.07.554.277.727.415-.1.52.035-.035.138-.1.519.208.485.311,1.211.1.415.069.346.312,1.142.276.173.381,1.315.138.242.484.519.035.277-.1.346-.415.415-.174.1-.346.1-.172-.069-.589-.554-.449-.277-.277-.138-.346-.035-.554.1-.934.45-.449.381c-2.338,3.155-1.5,5.255-.969,8.929h-.174l-.173-.727-.969.1-.727.1a5,5,0,0,1,.727,2.215h-.692l-.173-.554-.138-.173-1.35.45,1.35,2.907-.1.069-.761.381-.1.069-.138-.208a31.2,31.2,0,0,1-2.7,1.592l-.312-.519-2.007,1.211-1.038.623-1.939,1.211-.83.692c-1.723,1.6-1.561,3.087-2.631,5.433l.208,1.038.138.831.242,1.9.346,2.561-.761-.138-.485-.242-.138.138-.347-.035-.138.1h-.346l-.312-.138-.242-.208-1.107-.138-.381-.035-3.08-.242-.1.277"
-						transform="translate(19642.096 2957.171)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						className={
-							state.neighborhoodID === "43-horta"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="42-laclota"
-						d="M650.792,424.545l-.242.865.138,3.253.1,1.142-.207,1.488a9.942,9.942,0,0,1-.312,4.326l-5.295-1.938-4.88-.381-3.6-.554-2.388-.727-.519-.45-.554-.242-.97.069-.484-.208.519-1.765,1.039-.242,2.11-.658,4.638-1.9,2.908-2.872,4.914.554,2.007.035Z"
-						transform="translate(19657.93 2881.921)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "42-laclota"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="44-vilapicina"
-						d="M723.99,504.186c-3.757-4.329-7.46-7.595-10.279-12.563l-1.073.761-1.28,1.315-.172.173-.381.346-.035.311L708,497.472l-2.041-1.869-.624-1.315-3.219-7.2-.035-.069-1.28-2.872-.9-2.042-.208-.208-2.112-3.184-1.28-1.9c-.972-1.3-1.937-2.451-2.908-3.738L692.669,472l-.692-1-.45-.727-.242-.554.1-.277,3.08.242.381.035,1.107.138.242.208.312.138h.345l.138-.1.347.035.138-.138.484.242.761.138-.346-2.561-.242-1.9-.138-.831-.208-1.038.277-.692,1.142,2.734.8,1.8c9.06,9.588,18.238,19.1,27.34,28.655l-.242.554-1.28,4.5-.831,2.734Z"
-						transform="translate(19625.777 2860.594)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "44-vilapicina"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="45-porta"
-						d="M778.894,482c-6.522,3.916-12.871,8.106-19.485,11.871l-.934.173-1.176.623-.727.484-1.176,1-.658,1.107-.311,1.073-.139.346c-4.306-4.558-8.7-9.014-12.978-13.6,1.973-5.116,4.18-10.228,6.3-15.3,3.453-1.309,7.283-1.711,10.9-2.457C765.331,472.183,772.071,477.141,778.894,482Z"
-						transform="translate(19598.836 2858.453)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "45-porta"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="46-turodelapeira"
-						d="M722.285,475.182c-4.785-5.01-9.639-9.991-14.362-15.054l-.8-1.8-1.142-2.734.934-2.6.311-.831.554-.692.554-.623.83-.692,1.939-1.211.692,1.142.622.761.52.277.8.415.934.727.934.8c.853,1.488,2.211,3.963,2.283,5.641l1.73-.415.381,1.8,1.418-.277-.381-2.284,2.561-.934.346.692,1.9,2.007,2.734.588C726.465,464.953,724.258,470.065,722.285,475.182Z"
-						transform="translate(19617.859 2868.349)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "46-turodelapeira"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="47-canpeguera"
-						d="M717.086,441.274l.9-.554,2.145-1.281.312.519.588-.242.069-.035.1-.138.138-.035.277-.208.969-.623.554-.311.138.208.865-.45.1-.069-1.35-2.907,1.35-.45.207,2.146,1.834,2.63.969,1.973,1.661,2.8-.208,2.838.865,1.834-2.561.934.381,2.284-1.418.277-.381-1.8-1.73.415c-.23-2.244-.993-2.884-1.764-4.81l-.519-.831-.934-.8-.934-.727-.8-.415-.52-.277-.622-.761Z"
-						transform="translate(19611.883 2876.024)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "47-canpeguera"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="48-laguineueta"
-						d="M742.924,441.426l-2.734-.588-1.9-2.007-.346-.692-.865-1.834.208-2.838-1.661-2.8-.449-.934-.52-1.038-1.834-2.63-.1-1.038-.1-1.107.138.173.173.554.416.035.276-.035a4.668,4.668,0,0,0-.727-2.215l1.7-.208.173.727h.174l-.588-4.084v-.934l.415-1.8.381-1,.38-.45c7.816,3.445,16.247,5.619,24.4,8.271l-.8,2.6-.692,2.561L756.8,430.49l-.416,2.561-.034,2.284-.588,3.219C751.538,439.561,746.973,439.919,742.924,441.426Z"
-						transform="translate(19603.521 2886.808)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "48-laguineueta"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="49-canyelles"
-						d="M762.34,390.966c-7.837-2.6-15.892-4.579-23.43-7.96l.381-.658.449-.381.934-.45.554-.1.346.035.277.138.449.277.589.554.172.069.346-.1.174-.1.415-.415.1-.346-.035-.277-.484-.519-.138-.242-.381-1.315-.276-.173-.312-1.142-.069-.346-.1-.415-.311-1.211-.208-.484.1-.519.035-.138-.45-.035h-.138l-.346.1a2.9,2.9,0,0,1-.035-2.63,7.812,7.812,0,0,1,1.418-2.872l.242-.554.07-.277-.1-.658.208-.415.622-.865.139-.381.035-.934-.174-.381-.415-.623-.069-.623-.069-.8-.277-.865-.347-.519-.345-.381.069-.242.138-.173.172-.035.277.069.485.519,2.007,1.177.311.069,1.557-.727.346-.173.311-.208.277-.381.173-.45-.035-.346-.138-.588.1-.311.173-.1c1.533,1.167,1.342.761,3.53.623l1.454-.346,1.384-.311.935-.311.657-.173h.208c.671-.646.765-.377,1.246-1.038.657-1.237.093-.667.485-1.661l.138-.035.138.1.692.658.554.934.589.934,2.18,2.838,1.7,3.53.761.658,1.177,2.526.138,1.419-.312,2.042.52,2.042.276,2.734-.035.658-.242,1.073-.554,1.557-.588,1.938-1.661,4.568.588.311-.485,1.384Z"
-						transform="translate(19600.129 2918.476)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "49-canyelles"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="52-laprosperitat"
-						d="M798.97,471.819c-6.823-4.857-13.563-9.814-20.384-14.674,4.81-1.053,9.618-2.114,14.431-3.149,3.346-1.922,6.794-3.651,10.141-5.572l8.444,6.3,1.315.969,3.841,3.011,1.177,1.211-2.769,1.869-8.929,5.468-.622.692-3.946,2.284Z"
-						transform="translate(19578.762 2868.632)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "52-laprosperitat"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="51-verdun"
-						d="M805.419,439.9c-3.347,1.921-6.795,3.65-10.141,5.572-4.164.915-8.337,1.785-12.493,2.734l.588-3.219.034-2.284.416-2.561,1.627-2.388.691-2.561.8-2.6c5.671,1.561,10.911,3.658,16.4,5.745Z"
-						transform="translate(19576.498 2877.155)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "51-verdun"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="53-trinitatnova"
-						d="M846.613,447.392l-1.177-1.211-3.841-3.011-1.315-.969-8.444-6.3,3.288-.865.726-3.115.347-.831.692-1.35,3.045-.831.9-.485,1.7-.311,3.323.277,1.9-.311,1.419.069,2.18.035.831-1,1.142-.45,1.142.173.968-.346,1.558-.346.865-.588.069.035.519-.381.692-.1h.347l-.069.381,3.253.831,4.43-.173-.138,2.561-1.108,2.25-.934,1.35-1.523,1.973-1,1.246-1.072,1.315-.208.173-1.316,1.281-1.315,1.142-.311.277-1.177,1-2.042,1.523-1.591,1.073-.588.346-1.523.9-4.154,2.63Z"
-						transform="translate(19550.082 2881.154)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "53-trinitatnova"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="50-lesroquetes"
-						d="M790.654,410.545l-.968-.311,1.246-3.184.485-1.384-.588-.312c.749-3.28,2.757-6.4,3.079-9.794l1.212-1.765.9-1.834.8-.311,1.73.727.381.727-.554,2.319.277,1,.865.485,2.526.415,1.7.208.622,2.042,2.7.519.9-.346h1.488l1.454.45.172.381.485.865,1.142.969,2.837-.381,2.873,1.834,3.322.208,5.538,2.8,1.7-.208,1.73.208.969.484,1.834.381,1.315-.311.346.138-.865.588-1.558.346-.968.346-1.142-.173-1.142.45-.831,1-2.18-.035-1.42-.069-1.9.311-3.323-.277-1.7.311-.9.484-3.045.831-.692,1.35-.347.831-.726,3.115-3.288.865-2.076-1.557C801.663,414.18,796.236,412.116,790.654,410.545Z"
-						transform="translate(19572.783 2899.209)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "50-lesroquetes"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="54-torrebaro"
-						d="M800.991,359.627l1.558-3.461,4.049-4.291.208-.623.172-1.557.243-1.177.346-1.419,6.922,2.042,4.671,1.246,4.5,3.772,3.011.519,1.938,1.35.934.588.485,1.7,3.841,2.526,5.019.138,3.288.45,1.246.208,1.972.069.727.658.069.588,2.769.623-1.869,4.153.138,16.439-4.43.173-3.253-.831.069-.381h-.347l-.692.1-.519.381-.069-.035-.346-.138-1.315.311-1.834-.381-.969-.484-1.73-.208-1.7.208-5.538-2.8-3.322-.208-2.873-1.834-2.837.381-1.142-.969-.485-.865-.172-.381-1.454-.45h-1.488l-.9.346-2.7-.519-.622-2.042-1.7-.208-2.526-.415-.865-.484-.277-1,.554-2.319-.381-.727-1.73-.727-.8.311-.9,1.834-1.212,1.765-.276-2.734-.52-2.042.312-2.042-.138-1.419-1.177-2.526Z"
-						transform="translate(19569.957 2923.202)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "54-torrebaro"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="55-ciutatmeridiana"
-						d="M863.7,356.908l-2.769-.623-.069-.588-.727-.658-1.972-.069-1.246-.208-3.288-.45-5.019-.138-3.841-2.526-.485-1.7-.934-.588-1.937-1.35-3.011-.519-4.5-3.772-4.671-1.246-6.922-2.042,3.045-1.038.9-.311,2.7-1.419,2.8.485,2.388.346,1.626-.554.347.208,1.384.381-.38-.242,2.353-3.565,1,.658-1.73,2.63c1.393,1.214,1.2.5,2.769,1.281l.346.208.519.415.242.138.416.173h.1l.242.173.554.035.276.035.069.069.138.1.485.208.172.138.174.208.1.138.208.035.1.035.069.138.138.484.173.1.761-.173.208-.035.138.035.416-.277.172-.069h.069a16.607,16.607,0,0,0,2.388.311l.553.069.9.242.312.138,1.176-.138,2.25-.277,1.523.277.865.311,2.145,1,2.146.969.1.035,2.284,1.315c.529.29,1.038.573,1.558.9l-1.454,6.368-.831,3.7Z"
-						transform="translate(19555.213 2929.866)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "55-ciutatmeridiana"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="56-vallbona"
-						d="M908.86,382.341l-.9,2.769.138-2.561-.138-16.439,1.869-4.153.035-.069.831-3.7,1.454-6.368,1.177.658,1.245.727h.138l.138-.173.8-2.526.45-.138h.658l.173-.138.069.069.346-.208.069.035.588-.381.38-.1.347-.242-.277-.173-1.246.311-1.315-.208-.069-.035-.069-.1-.1-.138.277-1.454a9.372,9.372,0,0,1,2.976-3.5l.52,2.007c2.119,1.114,2.237,2.083,4.568,2.942l.727.519.139.173.519.311.174-.069.519.277c.3.347,1.415.8,1.9,1.315,1.24,1.417.42,1.589,1.212,3.6l1.073,1.107.208.242.277-.311.1.208,1.142.934.034.035q-7.725,8.623-15.4,17.269l-.934,1.177-1.039,1.384-.865,1.454-.449.761-.381.934-.415,1.107-.069.692-.242.934-1.177-.208-.311-.311-1-.484h-.8Z"
-						transform="translate(19509.084 2924.817)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "56-vallbona"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="14-lafontdelaguatlla"
-						d="M344.214,673.949l.035,3.945v2.63l.035,3.6-2.319-.554-2.976,1.9-1.834-.415-3.046-3.322-2.838-.484-1.246.658-4.707-1.142-1.523.45-.658,1.038.069.865-3.115-2.63-.277-.208-1.142.657-1.211-.346-1.973-1.9h-.554v-.831l1.834-.934,2.042-.969-.035-1.9C327.256,674.052,335.734,673.957,344.214,673.949Z"
-						transform="translate(19828.465 2747.173)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "14-lafontdelaguatlla"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="15-hostafrancs"
-						d="M359.644,648.906c-5.353,0-10.705.057-16.058.069l3.461-3.357,1.627-1.488,1.107-1.073-1.35-.346v-1.453l.208-.381,2.007-1.938-.761-.761,1.419-1.35.208-.1.484-.415.138-.173.069-.554.069-.208.485-.658-.588-.623,1.488-1.384,1.107-1.073,2.734,2.146,5.26-4.707,1.661-1.627,2.6.138c.036,6.578-.257,13.156-.035,19.727l-1.973-.1-1.454,1.661h-2.838Z"
-						transform="translate(19813.035 2772.216)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "15-hostafrancs"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="9-novaesquerradeleixample"
-						d="M423.487,624.538c-9.716.047-19.425.035-29.14.035-.956-3.989-.088-8.211-.346-12.286.239-3.776-.153-7.7.277-11.49a139.737,139.737,0,0,1,22.98-18.723l6.056,8.894.138,1.627v18.55Z"
-						transform="translate(19785.945 2796.653)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "9-novaesquerradeleixample"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="fp-zonafrancaport"
-						d="M239.935,729.45l-5.54,1.264-4.785,1.609-3.452,2.786-19.069,7.856-9.413,2.146-26.614.588-9.379-2.111-3.634-3.045-2.76-6-5.1-1.644-5.641-2.388-10.694-3.565-3.7-1.557-1.973-1.107-2.6-1.7-1.8-.865.6-5.48c-.953-.214-1.87-.656-2.826-.917l-1.378-.456.45-1.661-3.668-1.038-.208,1.557-1.973-.415a1.137,1.137,0,0,0-.727-1.142,11.878,11.878,0,0,1-3.249-.127,18.924,18.924,0,0,0-3.223-.323c.334-.565-.922-.353-.784-.81.337-.4-.8-.684-.878-.4l-1.384-.554c.2-.45-.687-.425-.18-1.068-.028-.366.592-1.173.08-1.208l-3.811-.423-.554,1.7a15.659,15.659,0,0,1-3.621-.977c-11.677-4.526-23.34-9.09-35.026-13.592-2.433-.913-4.9-1.756-7.312-2.713a25.339,25.339,0,0,0-3.7-1.491c-2.346-.883-4.755-1.6-7.079-2.541a9.909,9.909,0,0,0-2.816-1.032l-1.767-.427c3.161,15.409,2.049,29.567-3.046,44.506.4,2.145.745,4.306,1.176,6.44,1.468,3.133,2.874,6.3,4.381,9.411,2.3,4.09,4.629,8.162,7.055,12.179a68.11,68.11,0,0,0,4.044,5.791c1.278,1.652,2.5,3.358,3.812,4.977,4.87,4.428,9.779,8.818,14.625,13.269l15.158,17.7c2.085-.334,4.2-.594,6.265-.974,4.345-2.424,8.662-4.9,13.024-7.291.557.092.546-1.122.939-.539l7.5,7.641,39.983-19.7,1.227,1.76-6.229-9.448.727-2.769a8.2,8.2,0,0,1-2.6-1.63c-.954-.871-1.856-1.833-2.842-2.648l-4.318-1.19-6.541,2.492c-.241-.43-.41-1.052-.977-.608l-31.139,11.094-2.457-6.056c2.556-.856,4.962-2.092,7.444-3.134l1.45-.638-2.284-7.025,5.364-1.938,7.06,6.576,23.395-18.9,1.142,1.28-1.627,1.281q.37.9.98.142c.37-.222.805-.777,1.136-.8.127.419.852.935.252,1.237l-1.122,1.013c.464.426.494.792-.116,1.068l-2.2,1.735c.318.383.905.786.2,1.094l-3.246,2.817a12.107,12.107,0,0,0,3.485,1.193l20.152,3.029c.44-.45.366-1.188.585-1.762l.765-3.152c.525.356.866.262.982-.382.3-.673.414-1.446.739-2.093l2.362-2.059,1.557.381-4.153,18.965c.419-.01,1.129.44,1.353.121l4.531-18.81,2.769.554c-.149,1.121-.627,2.229-.392,3.367l1.292,12.276-1.523.554c.035.678.772.01,1.168,0,7.158-2.363,14.343-4.648,21.452-7.159l5.413-1.871-.519-9.621,4.984-1.869,3.7,11.005,6.472-2.284-2.331-8.109,11.732-3.149,2.607,8.49a21.653,21.653,0,0,0,3.039-.986c.944-.342,1.9-.658,2.844-.987.3-.422-.189-.922-.246-1.379-.234-.32-.137-1.317-.679-1.008-.61.459-.631-.3-.811-.734l-2.278-6.5c.386-.149,1.306-.113.643-.623-.371-.4.567-.638.77-.967l10.319-8.446c.16.627.581.683,1.129.461.461-.12-.124-.885-.111-1.282a1.45,1.45,0,0,1-.017-1.273C240.072,730,240,729.733,239.935,729.45Z"
-						transform="translate(19978.844 2740.589)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						className={
-							state.neighborhoodID === "fp-zonafrancaport"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="10-santantoni"
-						d="M424.095,688.732l-.1,4.084.035,1.073-.069,2.561-.069,2.146-.035,1.523v2.942l-.1.727-.069.8c-3.712-4-7.68-7.783-11.559-11.628-5.393-5.67-11.144-11.043-16.612-16.681l-.623-2.18c15.829.029,31.652-.056,47.482-.069l.069,2.042-2.423-.208C434.716,680.158,429.381,684.419,424.095,688.732Z"
-						transform="translate(19785.406 2747.133)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "10-santantoni"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="12-lamarinadelpratvermell"
-						d="M227.174,735.556l.658.069.242-3.288.415-2.63.208-.1,1-4.361-.381-.138.554-2.215.208-.45.138-.415,20.176,6.229-2.284,6.922,6.195,2.457,1.7.173-.554.727-.1,1.177,2.734,14.155-5.087-1.661-5.641-2.388-10.694-3.565-3.7-1.557-1.973-1.107-2.6-1.7-1.8-.865Z"
-						transform="translate(19876.047 2721.282)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "12-lamarinadelpratvermell"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="pm-parcdemontjuic"
-						d="M366.034,723.778l-3.513,2.769L343.453,734.4l-9.413,2.146-26.614.588-9.379-2.111-3.634-3.046-2.751-6.022-2.751-14.12.1-1.177.554-.727,1.35-2.63.519-.969-.138-1.627.173-1.765.761-1.246.277-.554,2.111-1.869c1.172-.094,3.861.328,4.81,1.108,1.157.672,2.126,3.283,3.253,4.188a1.675,1.675,0,0,0,1.869-.346,4.064,4.064,0,0,0,1.384-2.077c.763-1.434.65-4.49.485-6.126l-.138-.658-.277-.8-.346-1-.069-.831-.1-.45-.346-.311-.311-.415-.208-.519-.138-.519-.242-.8-.242-.692-.242-.865a24.519,24.519,0,0,0-.035-4.187l.831-.069c-.179-2.888-.337-2.57,1.523-5.4l3.115,2.63-.069-.865.658-1.038,1.523-.45,4.707,1.142,1.246-.658,2.838.484,3.046,3.322,1.834.415,2.976-1.9,2.319.554-.035-3.6v-2.63l-.035-3.945,1.073-.035h2.838l1.454-1.661,1.973.1.277,1.7.623,2.18c1.945,2.006,3.927,3.978,5.9,5.93l-1.366,3-2.353-.761-.865,2.7-1.765-.588-1.592,5.053-.381.865-.208.588v.415l.242.277.381.312.277.208.277.346.415.346.208.381.208.173.484.969,2.457,2.976,2.042-.519.069,1.869-1.177.035.831,2.7-.035.9-.865,2.872,1.35.381.969.588,1.419,1.177.658.657,6.956.035-.277,1,.208.761,1.073.277.865-.1,1.211.346.415.242.242.865.138.8.761.969.623.658,2.457,1.419,4.568,2.111-.623,1-1.142,1.627-.8,1.281Z"
-						transform="translate(19842.48 2749.152)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "pm-parcdemontjuic"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="11-poblesec"
-						d="M420.645,735.3l-3.219-2.353.8-1.281,1.142-1.627.623-1-4.568-2.111-2.457-1.419-.623-.658-.761-.969-.138-.8-.242-.865-.415-.242-1.211-.346-.865.1-1.073-.277-.208-.761.277-1-6.956-.035-.658-.658-1.419-1.177-.969-.588-1.35-.381.865-2.873.035-.9-.831-2.7,1.177-.035-.069-1.869-2.042.519-2.457-2.976-.484-.969-.208-.173-.208-.381-.415-.346-.277-.346-.277-.208-.381-.312-.242-.277v-.415l.208-.588.381-.865,1.592-5.053,1.765.588.865-2.7,2.353.761,1.35-3.011c5.474,5.316,10.736,10.774,16.162,16.127,2.492,2.592,4.9,4.7,7.2,7.475,3.914,2.971,6.8,7.839,10.14,11.455l1.28,2.6-3.322.8.035.173.485,2.111L425.49,733.7Z"
-						transform="translate(19787.75 2737.6)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "11-poblesec"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="69-diagonalmar"
-						d="M753.132,745.873l-.339,4.012c-.276-.233.035,3.349-.218,3.454-.718.176-.581-2.515-1.311-2.368-.764-.027-7.251-2.9-8.016-2.977-1.374.029-.916,2.6-1.062,3.508-1.013.131.283-5.294-.722-5.237l-1.034.058-.658-.1-.761.069-2.146-.1h-2.042l-.45.415-.9,3.149-.416-.242.692-2.838-.415-.588-.658-.346-.658-.069-1.869-.069-1.7.242-.449.173-.623.069-1.107.381-.692.069-.554.242-.865.208-.311.138-.761.208-1.593.519-1.141.45.035.658-.173.381-.035,3.322-.138.138-.174.1-.207-.1-.138-.138.035-2.353-.312-.519-.138-.415-.45-.277-1.316-.277-1.8-.1-2.18.138-1.038.208-3.08.554-1.938.588-.554.312-.172.242.1.277-.035.484-.242.727-.347.415-.035,2.319-.449-.069v-2.18l.242-1.488-.1-1.038h-.381l-.1-.519,4.326-1.627-.8-4.5.554-.138.034-2.63.589-.138-.243-5.226,3.184-.138.069-.069-.035-.381,7.268.415-.174-18.134v-1.765l3.427,1.73,4.707,2.111,2.907-.035,11.144-.1v6.783l10.9,5.157v1.488Z"
-						transform="translate(19618.266 2724.877)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "69-diagonalmar"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="70-besosimaresme"
-						d="M804.367,722.317a14.583,14.583,0,0,1,1.051,3.376l1.4,2.264-2.075,1.826c-.929-.034-5.848.67-6.758.821-.643.167.09-2.842-.542-2.73l-1.51.039.339-4.012-.035-13.393v-1.488l-10.9-5.156V681.99l11.178-2.734v-4.707h3.841l4.811.035h1.107l-.761,1.246v2.353c.943.922.681,2.346,1.073,3.5a20.4,20.4,0,0,1,.242,3.98c-.188,1.963.094,3.5,0,5.157-.355,1.9-.825,3.781-1.212,5.676a12.391,12.391,0,0,0,.139,3.392c.024,2.335-.432,2.961-1.038,5.087l-.1.831-.035.934.139.8.034.554-.1.381-.1.173-.1.1-1.349.242-.173.069-.035.069v.208l1.212,4.291Z"
-						transform="translate(19575.125 2746.85)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "70-besosimaresme"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="66-elparcilallaunadepoblenou"
-						d="M618.765,666.9a3.5,3.5,0,0,0-1.925.631,1.294,1.294,0,0,1-.852.351L612.293,666l-1.039.727,2.423,2.872h1.35c.053.256.274.568.234.783a5.837,5.837,0,0,0,.379.87,1.862,1.862,0,0,0,.667.874l-1,1-10.659,10.694-.692.415-2.423,2.526-2.042,1.834-10.175-.138-3.7-.069v7.233l3.668.035,3.6.069-.069,3.6,25.887.519.1-7.095,14.6-3.392.173-11.8.1-7.821-7.2-.035-.485.035h-.8l-2.457.035-.346-.588-1.073-1.384-1.592-.588Z"
-						transform="translate(19682.689 2751.455)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "66-elparcilallaunadepoblenou"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="5-elfortpienc"
-						d="M610.942,660.324l1-1a1.863,1.863,0,0,1-.667-.873,5.842,5.842,0,0,1-.379-.87c.04-.215-.182-.526-.234-.783h-1.35l-2.423-2.872c.346-.242.693-.484,1.025-.719-7.628-3.752-15.369-7.254-23-10.978l-.035,14.708-3.565-.035h-2.18l-.45,1.142-1.108.692-.035,2.353v3.772l-.035,3.634v3.7h3.772l-.035,3.392,3.7.069,10.175.138,2.042-1.834,2.423-2.526.692-.415Z"
-						transform="translate(19687.053 2764.259)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "5-elfortpienc"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="39-santgenisdelsagudells"
-						d="M520.375,360.563l-.069-.415-.588-.173-.623.415-.658.208-.45-.242-.277-.346-.138-.381.035-.381.242-.588v-.173l-.1-.138-.173-.1h-.761l-.277-.138-.173-.346-.242-.173-.554-.035-.658-.242-.484.069-.346.346.069.311.173.381.1.45-.1.831-.138.208-.208.138-.277.1h-.519l-.173.035-.138.138-.1.242v.484l.173,1.177-.1.969-.069.623-.173.069-.069-.242-.069.035-1.315-5.191-.9-3.53-.208-.692-.173-.519-.208-.45-2.25-8.1-.069.035-.173-.727-.069-.277.692.208.208.035.1-.035.346-.208.311-.554.173-.484.173-.969v-.554l.069-.381.138-.415.45-.9.312-.761.138-.311.173-.242.311-.173.242-.1.277-.277,1.107-1.419.485-.484.658-.381,3.6-1.315.588-.242.242-.208.173-.242.069-.277-.035-.311.208.035,2.526-.277.484-.277,1-.45.519-.138h.173l.311.069.242-.173.069-.035.658-.173.484-.346.311-.069.484-.173.138-.208.138-.069.208-.035,1.073-.554.381-.138,1.315.346,1-.242.415-.208,9.275-.485,1.419.381,1.35,1.107,1.765,2.457,1.661,1.038,1.281,2.6,2.526.658,1.281,2.872,1.246.069,1.246-1.973,1.073-1.107,2.872-1.523.658,1.177-1.142.8-1.869.381-1.073,1.973,1.454,11.836-4.949,2.146,1.765,3.634-2.284,1.627.173.831-1.142.623-1.246,3.288,2.492,4.776.1.173-1.973.934-6.818,3.253-2.319,1.142-2.353,1.177-1.177-2.63-.8-1.73-.208-.623-.035-.277.138-.277.346-.381.242-.138.381.035.934.277.346.069.346-.1.484-.346.277-.554c1.493-2.346,2.2-4.279.035-6.333l-.1-.588.242-.554,1.038-1.142.138-.208.035-.658-.069-.311-.415-1.523-.934.692-.727,1.28-2.942.277-1.661.969-.658,2.319-.554,1.592-.277.173v-1.523l-.069-.208L531,363.47l-.069-.1-.727-.519-.138-.1-.173-.1c-.8-.313-.871-1.129-1.453-1.246l-.138-.069-.692.035-.1.035h-.1c-.332-.079-.692-.624-1-.831l-.208-.173-.242-.138-.242-.1-.346-.242-.381-.242-.1-.1-.138-.069-.138-.1h-.069a2.551,2.551,0,0,0-.208.865l-.242,1.315-.138.485-.415,1.557-.1.311-.277.1-.208-.035-2.042-.692-.346-.1-.208-.1-.069-.1-.035-.242.035-.1.035-.277.173-.346.208-.623-.1-.623Z"
-						transform="translate(19724.98 2932.613)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "39-santgenisdelsagudells"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-					<path
-						id="25-santgervasibonanova"
-						d="M477.012,385.867l-.029.369.165.472c-.066,1.074,1.123.523,1.553.971-.2.643-1.119.789-1.375,1.429a4.914,4.914,0,0,0-.8,2.158,12.3,12.3,0,0,0,1.086,1.478,11.61,11.61,0,0,1,1.7.6l1.936,4.629-.085,3.982-.208,3.668-1.834,1.073-.381,3.876.485,1.246-.035.554-1.419,1.315-1.454,2.25-1.073,1.73-1.177,1.454-2.423,1.038-.311-.138-1.419-.519-.727,1.419-.727.173c-1.927,4.366-3.787,8.764-5.745,13.116l.069.692.277,3.011-.381-.138-1.557-.277-1.488-.311-4.6-.415c-4.527.649-9.129.64-13.636,1.454q2.345-9.2,4.811-18.377l-5.849-.865-8.652-1.315.485-.934.934-1.142.727-1.142.138-1.938v-.1l.173-2.25.934-2.18,1.765-2.353.969-4.084,2.734-.588.173-.658.554-1.28,1.038-1.627,2.111-1.834.069-.208v-.242l.035-.311v-.45l-.069-.45.035-.415.035-.346.208-.415V392.3l.277-.381.1-.415L445,389.466l.831-3.392.346-.138,1.211-1.038.208-.208.138-.415.138-.692.1-.381.519-.658.415-1.107.623-1.142.484-.346,1.384.554.727,1.384.346.138.658-.761.311-1.073.761-.761.485-.381.277.173.485,1.28.311.1.934-.381.277-.242,1.938-.277.173.242-.346.519-.311.173-.554.658v.692l.277.484-.45.519-.346.45-.1.485-.415.554-.519.623.069.311.588.069,1.177.069.1.311-.381.415-.173.312-.1.415-.554.519-1.35.761-.588.519.035.934.381.865.415.45,1.661.415.727.484.623.45.346-.069.1-.311.035-.588.9-1.661.727-.277,1.211-.381.415-.277.519.242,1.107.415,1.211-.554.865-.485.727.138.484.485.554,1.246,1.177,1.107.554.138,1.177-.588.934-.484.173-.069.069-.623.1-.969-.173-1.177v-.484l.1-.242.138-.138.173-.035h.519l.277-.1.208-.138.138-.208.1-.831-.1-.45-.173-.381-.069-.311.346-.346.484-.069.658.242.554.035.242.173.173.346.277.138h.761l.173.1.1.138v.173Z"
-						transform="translate(19765.574 2905.995)"
-						fill="none"
-						stroke="#000"
-						strokeLinecap="round"
-						strokeLinejoin="round"
-						strokeMiterlimit="10"
-						strokeWidth="0.1"
-						fillRule="evenodd"
-						className={
-							state.neighborhoodID === "25-santgervasibonanova"
-								? "hidden md:block lit-neighborhood"
-								: "hidden md:block"
-						}
-					/>
-
-					<text
-						id="2-gotic-nr"
-						transform="translate(20235.615 3451.573)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							2
-						</tspan>
-					</text>
-					<text
-						id="3-barceloneta-nr"
-						transform="translate(20254.227 3477.998)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							3
-						</tspan>
-					</text>
-					<text
-						id="4-santpere-nr"
-						transform="translate(20256.525 3453.412)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							4
-						</tspan>
-					</text>
-					<text
-						id="5-fortpienc-nr"
-						transform="translate(20276.746 3427.676)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							5
-						</tspan>
-					</text>
-					<text
-						id="6-sagradafamilia-nr"
-						transform="translate(20285.018 3401.941)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							6
-						</tspan>
-					</text>
-					<text
-						id="7-eixampledreta"
-						transform="translate(20247.104 3414.809)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							7
-						</tspan>
-					</text>
-					<text
-						id="8-antigaesquerradeleixample-nr"
-						transform="translate(20217.691 3405.158)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							8
-						</tspan>
-					</text>
-					<text
-						id="9-novaesquerradeleixample-nr"
-						transform="translate(20193.105 3407.226)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							9
-						</tspan>
-					</text>
-					<text
-						id="10-santantoni-nr"
-						transform="translate(20195.174 3432.042)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							10
-						</tspan>
-					</text>
-					<text
-						id="11-poblesec-nr"
-						transform="translate(20189.889 3452.033)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							11
-						</tspan>
-					</text>
-					<text
-						id="pm-parcmontjuic-nr"
-						transform="translate(20157.49 3466.739)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							PM
-						</tspan>
-					</text>
-					<text
-						id="fp-zonafranca-nr"
-						transform="translate(20057.996 3479.526)"
-						fontSize="8"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							FP
-						</tspan>
-					</text>
-					<text
-						id="12-lamarinadelpratvermell-nr"
-						transform="translate(20109.695 3460.764)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							12
-						</tspan>
-					</text>
-					<text
-						id="13-lamarinadelport"
-						transform="translate(20123.482 3440.314)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							13
-						</tspan>
-					</text>
-					<text
-						id="14-lafontdelaguatlla-nr"
-						transform="translate(20154.043 3427.676)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							14
-						</tspan>
-					</text>
-					<text
-						id="15-hostafrancs-nr"
-						transform="translate(20165.533 3415.268)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							15
-						</tspan>
-					</text>
-					<text
-						id="16-labordeta-nr"
-						transform="translate(20138.877 3415.958)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							16
-						</tspan>
-					</text>
-					<text
-						id="17-santsbadal-nr"
-						transform="translate(20136.58 3389.992)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							17
-						</tspan>
-					</text>
-					<text
-						id="18-sants-nr"
-						transform="translate(20154.963 3397.575)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							18
-						</tspan>
-					</text>
-					<text
-						id="19-lescorts-nr"
-						transform="translate(20174.033 3373.678)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							19
-						</tspan>
-					</text>
-					<text
-						id="20-lamaternitat-nr"
-						transform="translate(20133.594 3358.742)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							20
-						</tspan>
-					</text>
-					<text
-						id="21-pedralbes"
-						transform="translate(20141.865 3324.275)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							21
-						</tspan>
-					</text>
-					<text
-						id="22-vallvidrera-nr"
-						transform="translate(20180.008 3236.04)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							22
-						</tspan>
-					</text>
-					<text
-						id="23-sarria-nr"
-						transform="translate(20179.549 3308.88)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							23
-						</tspan>
-					</text>
-					<text
-						id="24-trestorres"
-						transform="translate(20194.715 3340.59)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							24
-						</tspan>
-					</text>
-					<text
-						id="25-santgervasibonanova-nr"
-						transform="translate(20216.314 3317.152)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							25
-						</tspan>
-					</text>
-					<text
-						id="26-santgervasigalvany-nr"
-						transform="translate(20210.34 3363.568)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							26
-						</tspan>
-					</text>
-					<text
-						id="27-elputget-nr"
-						transform="translate(20233.547 3345.645)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							27
-						</tspan>
-					</text>
-					<text
-						id="28-vallcarca-nr"
-						transform="translate(20249.172 3318.99)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							28
-						</tspan>
-					</text>
-					<text
-						id="29-elcoll-nr"
-						transform="translate(20265.027 3328.182)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							29
-						</tspan>
-					</text>
-					<text
-						id="30-lasalut-nr"
-						transform="translate(20261.58 3350.7)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							30
-						</tspan>
-					</text>
-					<text
-						id="31-gracia"
-						transform="translate(20246.645 3375.057)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							31
-						</tspan>
-					</text>
-					<text
-						id="32-elcampdengrassot-nr"
-						transform="translate(20265.717 3387.924)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							32
-						</tspan>
-					</text>
-					<text
-						id="33-baixguinardo-nr"
-						transform="translate(20283.18 3375.287)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							33
-						</tspan>
-					</text>
-					<text
-						id="34-canbaro-nr"
-						transform="translate(20284.328 3356.215)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							34
-						</tspan>
-					</text>
-					<text
-						id="35-elguinardo-nr"
-						transform="translate(20308.455 3370.461)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							35
-						</tspan>
-					</text>
-					<text
-						id="36-lafontdenfargues"
-						transform="translate(20307.076 3344.036)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							36
-						</tspan>
-					</text>
-					<text
-						id="37-elcarmel"
-						transform="translate(20281.342 3335.764)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							37
-						</tspan>
-					</text>
-					<text
-						id="38-lateixonera"
-						transform="translate(20268.703 3313.476)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							38
-						</tspan>
-					</text>
-					<text
-						id="39-santgenisdelsagudells-nr"
-						transform="translate(20251.471 3281.996)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							39
-						</tspan>
-					</text>
-					<text
-						id="40-montbau-nr"
-						transform="translate(20289.154 3271.656)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							40
-						</tspan>
-					</text>
-					<text
-						id="41-valldhebron-nr"
-						transform="translate(20291.451 3303.136)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							41
-						</tspan>
-					</text>
-					<text
-						id="42-laclota-nr"
-						transform="translate(20296.506 3313.476)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							42
-						</tspan>
-					</text>
-					<text
-						id="43-horta-nr"
-						transform="translate(20322.012 3284.294)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							43
-						</tspan>
-					</text>
-					<text
-						id="44-vilapicina-nr"
-						transform="translate(20328.676 3348.862)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							44
-						</tspan>
-					</text>
-					<text
-						id="45-porta-nr"
-						transform="translate(20349.586 3341.739)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							45
-						</tspan>
-					</text>
-					<text
-						id="46-turodelapeira-nr"
-						transform="translate(20331.434 3334.845)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							46
-						</tspan>
-					</text>
-					<text
-						id="47-canpeguera-nr"
-						transform="translate(20330.514 3322.897)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							47
-						</tspan>
-					</text>
-					<text
-						id="48-laguineueta-nr"
-						transform="translate(20345.449 3317.612)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							48
-						</tspan>
-					</text>
-					<text
-						id="49-canyelles-nr"
-						transform="translate(20348.438 3295.323)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							49
-						</tspan>
-					</text>
-					<text
-						id="50-lesroquetes-nr"
-						transform="translate(20370.725 3307.961)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							50
-						</tspan>
-					</text>
-					<text
-						id="51-verdun-nr"
-						transform="translate(20362.453 3319.91)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							51
-						</tspan>
-					</text>
-					<text
-						id="52-prosperitat-nr"
-						transform="translate(20373.482 3330.479)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							52
-						</tspan>
-					</text>
-					<text
-						id="53-trinitatnova-nr"
-						transform="translate(20392.785 3318.99)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							53
-						</tspan>
-					</text>
-					<text
-						id="54-torrebaro-nr"
-						transform="translate(20387.039 3293.945)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							54
-						</tspan>
-					</text>
-					<text
-						id="55-ciutatmeridiana-nr"
-						transform="translate(20402.436 3280.158)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							55
-						</tspan>
-					</text>
-					<text
-						id="56-vallbona-nr"
-						transform="translate(20421.736 3287.97)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							56
-						</tspan>
-					</text>
-					<text
-						id="57-trinitatvella-nr"
-						transform="translate(20409.098 3331.398)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							57
-						</tspan>
-					</text>
-					<text
-						id="58-barodeviver-nr"
-						transform="translate(20409.787 3350.47)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							58
-						</tspan>
-					</text>
-					<text
-						id="59-bonpastor-nr"
-						transform="translate(20393.703 3375.057)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							59
-						</tspan>
-					</text>
-					<text
-						id="60-santandreu-nr"
-						transform="translate(20372.334 3354.606)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							60
-						</tspan>
-					</text>
-					<text
-						id="61-sagrera-nr"
-						transform="translate(20345.221 3385.397)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							61
-						</tspan>
-					</text>
-					<text
-						id="62-elcongresielsindians-nr"
-						transform="translate(20332.123 3367.704)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							62
-						</tspan>
-					</text>
-					<text
-						id="63-navas-nr"
-						transform="translate(20327.527 3390.222)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							63
-						</tspan>
-					</text>
-					<text
-						id="64-campdelarpadelclot-nr"
-						transform="translate(20307.535 3398.035)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							64
-						</tspan>
-					</text>
-					<text
-						id="65-elclot-nr"
-						transform="translate(20316.957 3415.268)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							65
-						</tspan>
-					</text>
-					<text
-						id="66-elparcilallacunadelpoblenou-nr"
-						transform="translate(20291.451 3440.084)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							66
-						</tspan>
-					</text>
-					<text
-						id="67-vilaolimpica-nr"
-						transform="translate(20283.869 3464.671)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							67
-						</tspan>
-					</text>
-					<text
-						id="68-poblenou-nr"
-						transform="translate(20312.592 3455.709)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							68
-						</tspan>
-					</text>
-					<text
-						id="69-diagonalmar-nr"
-						transform="translate(20345.449 3462.833)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							69
-						</tspan>
-					</text>
-					<text
-						id="70-besosimaresme-nr"
-						transform="translate(20365.441 3442.612)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							70
-						</tspan>
-					</text>
-					<text
-						id="71-provençalsdelpoblenou-nr"
-						transform="translate(20339.475 3433.421)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							71
-						</tspan>
-					</text>
-					<text
-						id="72-santmartideprovençals-nr"
-						transform="translate(20342.463 3413.2)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							72
-						</tspan>
-					</text>
-					<text
-						id="73-lavernedailapau-nr"
-						transform="translate(20365.441 3406.307)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							73
-						</tspan>
-					</text>
-					<text
-						id="1-elraval-nr"
-						transform="translate(20218.842 3444.91)"
-						fontSize="9"
-						fontFamily="ArialMT, Arial"
-					>
-						<tspan x="0" y="0">
-							1
-						</tspan>
-					</text>
-				</g>
-			</SvgStyled>
-		</div>
-	);
-};
-
-export default InteractiveMap;
