--- frontend/src/components/organisms/interactive-map/data/all-districts.js
@@ -1,435 +1,0 @@
-export const allDistricts = [
-	{
-		district: "Ciutat-Vella",
-		neighborhoods: [
-			{
-				nr: 1,
-				name: "el Raval",
-				id: "1-raval",
-			},
-			{
-				nr: 2,
-				name: "el Gòtic",
-				id: "2-gotic",
-			},
-			{
-				nr: 3,
-				name: "la Barceloneta",
-				id: "3-barceloneta",
-			},
-			{
-				nr: 4,
-				name: "Sant Pere, Santa Caterina i la Rivera",
-				id: "4-santpere",
-			},
-		],
-	},
-	{
-		district: "L-Eixample",
-		neighborhoods: [
-			{
-				nr: 5,
-				name: "el Fort Pienc",
-				id: "5-elfortpienc",
-			},
-			{
-				nr: 6,
-				name: "la Sagrada Familia",
-				id: "6-sagradafamilia",
-			},
-			{
-				nr: 7,
-				name: "la Dreta de l'Eixample",
-				id: "7-eixampledreta",
-			},
-			{
-				nr: 8,
-				name: "Antiga esquerra de l'Eixample",
-				id: "8-antigadretadeleixample",
-			},
-			{
-				nr: 9,
-				name: "Nova esquerra de l'Eixample",
-				id: "9-novaesquerradeleixample",
-			},
-			{
-				nr: 10,
-				name: "Sant Antoni",
-				id: "10-santantoni",
-			},
-		],
-	},
-	{
-		district: "Sants-Montjuic",
-		neighborhoods: [
-			{
-				nr: 11,
-				name: "el Poble Sec",
-				id: "11-poblesec",
-			},
-			{
-				nr: 12,
-				name: "la Marina del Prat Vermell",
-				id: "12-lamarinadelpratvermell",
-			},
-			{
-				nr: 13,
-				name: "la Marina de Port",
-				id: "13-marinadeport",
-			},
-			{
-				nr: 14,
-				name: "la Font de la Guatlla",
-				id: "14-lafontdelaguatlla",
-			},
-			{
-				nr: 15,
-				name: "Hostafrancs",
-				id: "15-hostafrancs",
-			},
-			{
-				nr: 16,
-				name: "la Bordeta",
-				id: "16-labordeta",
-			},
-			{
-				nr: 17,
-				name: "Sants-Badal",
-				id: "17-santsbadal",
-			},
-			{
-				nr: 18,
-				name: "Sants",
-				id: "18-sants",
-			},
-			{
-				nr: "PM",
-				name: "Parc de Montjuic",
-				id: "pm-parcdemontjuic",
-			},
-			{
-				nr: "ZF",
-				name: "Zona Franca-Port",
-				id: "fp-zonafrancaport",
-			},
-		],
-	},
-	{
-		district: "Les-Corts",
-		neighborhoods: [
-			{
-				nr: 19,
-				name: "les Corts",
-				id: "19-lescorts",
-			},
-			{
-				nr: 20,
-				name: "la Maternitat i Sant Ramon",
-				id: "20-lamaternitat",
-			},
-			{
-				nr: 21,
-				name: "Pedralbes",
-				id: "21-pedralbes",
-			},
-		],
-	},
-	{
-		district: "Sarria-Sant-Gervasi",
-		neighborhoods: [
-			{
-				nr: 22,
-				name: "Vallvidrera, Tibidabo i les Planes",
-				id: "22-vallvidrera",
-			},
-			{
-				nr: 23,
-				name: "Sarrià",
-				id: "23-sarria",
-			},
-			{
-				nr: 24,
-				name: "les Tres Torres",
-				id: "24-trestorres",
-			},
-			{
-				nr: 25,
-				name: "Sant Gervasi-Bonanova",
-				id: "25-santgervasibonanova",
-			},
-			{
-				nr: 26,
-				name: "Sant Gervasi-Galvany",
-				id: "26-santgervasi",
-			},
-			{
-				nr: 27,
-				name: "el Putget i Farró",
-				id: "27-elputget",
-			},
-			{
-				nr: null,
-				id: "path3515",
-			},
-			{
-				nr: null,
-				id: "path3509",
-			},
-		],
-	},
-	{
-		district: "Gracia",
-		neighborhoods: [
-			{
-				nr: 28,
-				name: "Vallcarca i els Penitents",
-				id: "28-vallcarca",
-			},
-			{
-				nr: 29,
-				name: "el Coll",
-				id: "29-elcoll",
-			},
-			{
-				nr: 30,
-				name: "la Salut",
-				id: "30-lasalut",
-			},
-			{
-				nr: 31,
-				name: "Vila de Gràcia",
-				id: "31-gracia",
-			},
-			{
-				nr: 32,
-				name: "el Camp d'en Grassot i Gràcia Nova",
-				id: "32-elcampdengrassot",
-			},
-		],
-	},
-	{
-		district: "Horta-Guinardo",
-		neighborhoods: [
-			{
-				nr: 33,
-				name: "Baix Guinardó",
-				id: "33-baixguinardo",
-			},
-			{
-				nr: 34,
-				name: "Can Baró",
-				id: "34-canbaro",
-			},
-			{
-				nr: 35,
-				name: "el Guinardó",
-				id: "35-elguinardo",
-			},
-			{
-				nr: 36,
-				name: "la Font d'en Fargues",
-				id: "36-lafontdenfargues",
-			},
-			{
-				nr: 37,
-				name: "el Carmel",
-				id: "37-elcarmel",
-			},
-			{
-				nr: 38,
-				name: "la Teixonera",
-				id: "38-lateixonera",
-			},
-			{
-				nr: 39,
-				name: "Sant Genís dels Agudells",
-				id: "39-santgenisdelsagudells",
-			},
-			{
-				nr: 40,
-				name: "Montbau",
-				id: "40-montbau",
-			},
-			{
-				nr: 41,
-				name: "la Vall d'Hebron",
-				id: "41-valldhebron",
-			},
-			{
-				nr: 42,
-				name: "la Clota",
-				id: "42-laclota",
-			},
-			{
-				nr: 43,
-				name: "Horta",
-				id: "43-horta",
-			},
-		],
-	},
-	{
-		district: "Nou-Barris",
-		neighborhoods: [
-			{
-				nr: 44,
-				name: "Vilapicina-Torre Llobeta",
-				id: "44-vilapicina",
-			},
-			{
-				nr: 45,
-				name: "Porta",
-				id: "45-porta",
-			},
-			{
-				nr: 46,
-				name: "el Turó de la Peira",
-				id: "46-turodelapeira",
-			},
-			{
-				nr: 47,
-				name: "Can Peguera",
-				id: "47-canpeguera",
-			},
-			{
-				nr: 48,
-				name: "la Guineueta",
-				id: "48-laguineueta",
-			},
-			{
-				nr: 49,
-				name: "Canyelles",
-				id: "49-canyelles",
-			},
-			{
-				nr: 50,
-				name: "les Roquetes",
-				id: "50-lesroquetes",
-			},
-			{
-				nr: 51,
-				name: "Verdun",
-				id: "51-verdun",
-			},
-			{
-				nr: 52,
-				name: "la prosperitat",
-				id: "52-laprosperitat",
-			},
-			{
-				nr: 53,
-				name: "la Trinitat Nova",
-				id: "53-trinitatnova",
-			},
-			{
-				nr: 54,
-				name: "Torre Baró",
-				id: "54-torrebaro",
-			},
-			{
-				nr: 55,
-				name: "Ciutat Meridiana",
-				id: "55-ciutatmeridiana",
-			},
-			{
-				nr: 56,
-				name: "Vallbona",
-				id: "56-vallbona",
-			},
-		],
-	},
-	{
-		district: "Sant-Andreu",
-		neighborhoods: [
-			{
-				nr: 57,
-				name: "la Trinitat Vella",
-				id: "57-trinitatvella",
-			},
-			{
-				nr: 58,
-				name: "Baró de Viver",
-				id: "58-barodeviver",
-			},
-			{
-				nr: 59,
-				name: "el Bon Pastor",
-				id: "59-bonpastor",
-			},
-			{
-				nr: 60,
-				name: "Sant Andreu",
-				id: "60-santandreu",
-			},
-			{
-				nr: 61,
-				name: "la Sagrera",
-				id: "61-lasagrera",
-			},
-			{
-				nr: 62,
-				name: "el Congrès i els Indians",
-				id: "62-elcongresielsindians",
-			},
-			{
-				nr: 63,
-				name: "Navas",
-				id: "63-navas",
-			},
-		],
-	},
-	{
-		district: "Sant-Marti",
-		neighborhoods: [
-			{
-				nr: 64,
-				name: "el Camp de l'Arpa del Clot",
-				id: "64-campdelarpa",
-			},
-			{
-				nr: 65,
-				name: "el Clot",
-				id: "65-elclot",
-			},
-			{
-				nr: 66,
-				name: "el Parc i la Llacuna del Poblenou",
-				id: "66-elparcilallaunadepoblenou",
-			},
-			{
-				nr: 67,
-				name: "la Vila Olímpica del  Poblenou",
-				id: "67-vilaolimpica",
-			},
-			{
-				nr: 68,
-				name: "el Poblenou",
-				id: "68-poblenou",
-			},
-			{
-				nr: 69,
-				name: "Diagonal Mar i el Front Marítim del Poblenou",
-				id: "69-diagonalmar",
-			},
-			{
-				nr: 70,
-				name: "el Besòs i el Maresme",
-				id: "70-besosimaresme",
-			},
-			{
-				nr: 71,
-				name: "Provençals del Poblenou",
-				id: "71-provençalsdelpoblenou",
-			},
-			{
-				nr: 72,
-				name: "Sant Martí de Provençals",
-				id: "72-santmartideprovençals",
-			},
-			{
-				nr: 73,
-				name: "la Verneda i la pau",
-				id: "73-lavernedailapau",
-			},
-		],
-	},
-];
