--- frontend/src/components/organisms/interactive-map/store/context.js
@@ -1,3 +1,0 @@
-import {createContext} from "react";
-
-export const MapContext = createContext(null);
